### 2022-11-27 #4972 @editorialbot
[PRE REVIEW]: Persistable: persistent and stable clustering
**Submitting author:** @LuisScoccola (<a href="http://orcid.org/0000-0002-4862-722X">Luis Scoccola</a>)
**Repository:** https://github.com/LuisScoccola/persistable
**Branch with paper.md** (empty if default branch): https://github.com/LuisScoccola/persistable/tree/paper
**Version:** 0.3.2
