### 2022-11-27 #4972 @LuisScoccola
**Potential reviewers:** contributors to the [ripser python package (published at JOSS)](https://github.com/scikit-tda/ripser.py) (e.g., _ctralie_, _sauln_), the [hdbscan python package (published at JOSS)](https://github.com/scikit-learn-contrib/hdbscan) (e.g., _lmcinnes_, _sastels_, _jc-healy_), and [rivet](https://github.com/rivetTDA/rivet) (e.g., _xoltar_, _mlwright84_) may be suitable because of their expertise in computational persistence; although only _sauln_ and _lmcinnes_ are current reviewers for JOSS. As for current JOSS reviewers, besides the ones already mentioned, the ones mentioning topological data analysis (_ismailguzel_, _leotrs_, _EduPH_, _rrrlw_, _corybrunson_) are probably suitable, as are the ones mentioning clustering (_amarkos_, _gagolews_, _jonjoncardoso_, _TahiriNadia_, _AP6YC_).

**Lack of statement of need:** we have a section named "Related work and statement of need", please let us know if this should be adjusted in any way.
