### 2022-12-16 #4972 @adavidzh
> 👋 @adavidzh and @lmcinnes, would you be willing to review this submission for JOSS?

I'm afraid that I cannot take this on at this time.
