### 2022-12-16 4e77384
add betti number computation
