+++ persistable/persistable_interactive.py
@@ -8,7 +8,8 @@ import plotly.graph_objects as go
 import plotly
 from plotly.colors import sample_colorscale
 import json
-#import diskcache
+
+# import diskcache
 import dash
 from dash import dcc, html, DiskcacheManager, ctx
 from dash.exceptions import PreventUpdate
@@ -16,6 +17,7 @@ from jupyter_dash import JupyterDash
 import click
 import socket
 from ._vineyard import Vineyard
+from .signed_betti_numbers import signed_betti
 
 import threading
 
@@ -48,6 +50,7 @@ DISPLAY_LINES_SELECTION = "display-lines-selection-"
 ENDPOINT_SELECTION = "endpoint-selection-"
 STORED_CCF = "stored-ccf-"
 STORED_CCF_DRAWING = "stored-ccf-drawing-"
+STORED_BETTI = "stored-betti-"
 MIN_DIST_SCALE = "min-dist-scale-"
 MAX_DIST_SCALE = "max-dist-scale-"
 MIN_DENSITY_THRESHOLD = "min-density-threshold-"
@@ -60,6 +63,7 @@ STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
 INPUT_GRANULARITY_CCF = "input-granularity-ccf-"
 INPUT_NUM_JOBS_CCF = "input-num-jobs-ccf-"
 INPUT_MAX_COMPONENTS = "input-max-components-"
+INPUT_SIGNED_BETTI_NUMBERS = "input-signed-betti-numbers-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
 PV_DETAILS = "pv-details-"
@@ -134,7 +138,6 @@ class PersistableInteractive:
         self._parameters_sem = threading.Semaphore()
         self._parameters = None
 
-
     def start_UI(self, port=8050, debug=False, inline=False):
         """Serves the GUI with a given persistable instance.
 
@@ -157,7 +160,7 @@ class PersistableInteractive:
         if debug:
             self._debug = debug
         max_port = 65535
-        for possible_port in range(port, max_port+1):
+        for possible_port in range(port, max_port + 1):
             # check if port is in use
             with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
                 in_use = s.connect_ex(("localhost", int(possible_port))) == 0
@@ -167,9 +170,7 @@ class PersistableInteractive:
                 port = possible_port
                 break
         if possible_port == max_port:
-            raise Exception(
-                "All ports are already in use. Cannot start the GUI."
-            )
+            raise Exception("All ports are already in use. Cannot start the GUI.")
 
         background_callback_manager = DiskcacheManager()
 
@@ -178,20 +179,20 @@ class PersistableInteractive:
             self._app = JupyterDash(
                 __name__,
                 background_callback_manager=background_callback_manager,
-                update_title = "Persistable is computing..."
+                update_title="Persistable is computing...",
             )
             self._layout_gui()
             self._register_callbacks(self._persistable, self._debug)
 
-            #import logging
-            #self._app.logger.setLevel(logging.WARNING)
-            #logging.getLogger("werkzeug").setLevel(logging.ERROR)
+            # import logging
+            # self._app.logger.setLevel(logging.WARNING)
+            # logging.getLogger("werkzeug").setLevel(logging.ERROR)
             self._app.run_server(port=port, mode="inline", debug=debug)
         else:
             self._app = dash.Dash(
                 __name__,
                 background_callback_manager=background_callback_manager,
-                update_title = "Persistable is computing..."
+                update_title="Persistable is computing...",
             )
[…]
