### 2022-12-17 64b1c1a
improve layout of signed betti numbers
