### 2022-12-19 #5022 @editorialbot
[REVIEW]: Persistable: persistent and stable clustering
**Submitting author:** @LuisScoccola (<a href="http://orcid.org/0000-0002-4862-722X">Luis Scoccola</a>)
**Repository:** https://github.com/LuisScoccola/persistable
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.4.0
