### 2022-12-20 #5022 @rkurchin
@LuisScoccola, it looks like you may have a typo in a citation somewhere that's borking the BibTeX checker, take a look at that when you can since those checks will eventually have to pass to accept the submission.
