### 2022-12-21 8e3d5c5
improve tex bibliography formatting
