+++ paper/paper.bib
@@ -50,13 +50,16 @@
 }
 
 @article{mcinneshealyastels,
-  title={hdbscan: Hierarchical density based clustering.},
-  author={McInnes, Leland and Healy, John and Astels, Steve},
-  journal={Journal of Open Source Software},
-  volume={2},
-  number={11},
-  pages={205},
-  year={2017}
+    doi = {10.21105/joss.00205},
+    url = {https://doi.org/10.21105/joss.00205},
+    year = {2017},
+    publisher = {The Open Journal},
+    volume = {2},
+    number = {11},
+    pages = {205},
+    author = {Leland McInnes and John Healy and Steve Astels},
+    title = {hdbscan: Hierarchical density based clustering},
+    journal = {Journal of Open Source Software}
 }
 
 @online{plotly,
