### 2022-12-22 #5022 @LuisScoccola
@rkurchin Thank you for pointing out the bibtex issue. I think I have fixed it. Please let me know if there is anything that I should look into.
