### 2022-12-26 bff6925
add hierarchical clustering over vertical slice
