### 2022-12-26 b22a240
start rank invariant computation
