+++ persistable/persistable.py
@@ -363,7 +363,7 @@ class _MetricProbabilitySpace:
         self._kernel_estimate = None
         self._n_neighbors = None
         self._maxs = None
-        self._tol = _TOL
+        #self._tol = _TOL
         if metric in KDTree.valid_metrics:
             self._tree = KDTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
         elif metric in BallTree.valid_metrics:
@@ -532,8 +532,7 @@ class _MetricProbabilitySpace:
             )
         return np.array(density_estimates)
 
-
-    def lambda_linkage_vertical(self, s_intercept, k_start, k_end):
+    def _lambda_linkage_vertical(self, s_intercept, k_start, k_end):
         if k_end >= k_start:
             raise ValueError("Parameters do not give a monotonic line.")
 
@@ -587,13 +586,9 @@ class _MetricProbabilitySpace:
             core_distances, merges, merges_heights, hc_start, hc_end
         )
 
-
-    def lambda_linkage(self, start, end):
-        if start[0] > end[0] or start[1] < end[1]:
-            raise ValueError("Parameters do not give a monotonic line.")
-
+    def _lambda_linkage_skew(self, start, end):
         def _startend_to_intercepts(start, end):
-            if end[0] == np.infty:
+            if end[0] == np.infty or start[1] == end[1]:
                 k_intercept = start[1]
                 s_intercept = np.infty
             else:
@@ -653,6 +648,18 @@ class _MetricProbabilitySpace:
             core_distances, merges, merges_heights, hc_start, hc_end
         )
 
+    def lambda_linkage(self, start, end):
+        if start[0] > end[0] or start[1] < end[1]:
+            raise ValueError("Parameters do not give a monotonic line.")
+
+        if start[0] == end[0]:
+            s_intercept = start[0]
+            k_start = start[1]
+            k_end = end[1]
+            return self._lambda_linkage_vertical(s_intercept, k_start, k_end)
+        else:
+            return self._lambda_linkage_skew(start,end)
+
     def lambda_linkage_vineyard(self, startends, n_jobs, tol=_TOL):
         run_in_parallel = lambda startend: self.lambda_linkage(
             startend[0], startend[1]
@@ -675,14 +682,94 @@ class _MetricProbabilitySpace:
                     delayed(run_in_parallel)(startend) for startend in startends
                 )
 
+    def _lambda_linkage_parallel(self, startends, n_jobs):
+        run_in_parallel = lambda startend: self.lambda_linkage(
+            startend[0], startend[1]
+        )
+        if n_jobs == 1:
+            return [run_in_parallel(startend) for startend in startends]
+        else:
+            verbose = 11 if self._debug else 0
+            n_jobs = min(cpu_count(), n_jobs)
+            if self._threading:
+                return Parallel(n_jobs=n_jobs, backend="threading", verbose=verbose)(
+                    delayed(run_in_parallel)(startend) for startend in startends
+                )
+            else:
+                return Parallel(n_jobs=n_jobs, verbose=verbose)(
+                    delayed(run_in_parallel)(startend) for startend in startends
+                )
 
-
-    def hilbert_function(self, ks, ss, n_jobs, tol=_TOL):
+    def rank_invariant(self, ks, ss, n_jobs):
         n_s = len(ss)
         n_k = len(ks)
-        tol = ss[1] - ss[0]
-        startends = [[[0, k], [np.infty, k]] for k in ks[:-1]]
[…]
