### 2022-12-26 aea36c4
try slow version of rank invariant
