+++ persistable/persistable.py
@@ -533,7 +533,7 @@ class _MetricProbabilitySpace:
         return np.array(density_estimates)
 
     def _lambda_linkage_vertical(self, s_intercept, k_start, k_end):
-        if k_end >= k_start:
+        if k_end > k_start:
             raise ValueError("Parameters do not give a monotonic line.")
 
         indices = np.arange(self._size)
@@ -700,69 +700,102 @@ class _MetricProbabilitySpace:
                     delayed(run_in_parallel)(startend) for startend in startends
                 )
 
-    def rank_invariant(self, ks, ss, n_jobs):
-        n_s = len(ss)
-        n_k = len(ks)
-        #tol = ss[1] - ss[0]
-        startends_horizontal = [ [[ss[0], k], [ss[-1], k]] for k in ks[:-1] ]
-        startends_vertical = [ [[s, ks[0]], [s, ks[-1]]] for s in ss[:-1] ]
-        startends = startends_horizontal + startends_vertical
-        hcs = self._lambda_linkage_parallel(startends, n_jobs=n_jobs)
-        hcs_horizontal = hcs[:n_k]
-        hcs_vertical = hcs[n_k+1:]
+    #def rank_invariant_slow(self, ks, ss):
+    #    n_s = len(ss)
+    #    n_k = len(ks)
+    #    #tol = ss[1] - ss[0]
+    #    hf = self.hilbert_function(ks, ss, 1):
+    #    ri = np.zeros((n_s, n_k, n_s, n_k))
+    #    for i in range(n_s):
+    #        for j in range(n_k):
+    #            ri[i,j,i,j] = hf[]
+    #    for i in range(n_s):
+    #        for j in range(n_k):
+    #            for i_ in range(i+1, n_s):
+    #                for j_ in range(j):
+    #                    print([ss[i],ks[j]], [ss[i_],ks[j_]])
+    #                    hc = self.lambda_linkage([ss[i],ks[j]], [ss[i_],ks[j_]])
+    #                    l = hc._end - hc._start
+    #                    pd = hc.persistence_diagram(tol=0)
+    #                    print(pd)
+    #                    proms = pd[:,1] - pd[:,0]
+    #                    print(proms)
+    #                    rk = np.count_nonzero(proms == l)
+    #                    ri[i,j,i_,j_] = rk
+    #    print(ri)
+
+    #    #for i, pd in enumerate(pds):
+    #    #    for bar in pd:
+    #    #        b, d = bar
+    #    #        start = np.searchsorted(ss[:-1], b)
+    #    #        end = np.searchsorted(ss[:-1], d)
+    #    #        hf[i, start:end] += 1
+    #    #return hf
 
 
-        def _splice_hcs(s_index,k_index):
-            # the vertical hierarchical clustering
-            ver_hc = hcs_vertical[k_index]
-            n_ver_merges = ver_merges.shape[0]
-            ## push all things that happened before ks[k_index] there, and index starting from ss[s_index]
-            #ver_heights = ss[s_index] + np.minimum(ks[k_index], ver_hc._heights)
-            ver_heights = ss[s_index] + ver_hc._heights[ver_hc._heights < ks[k_index]]
-            ## push all things that happened before ks[k_index] there, and index starting from ss[s_index]
-            #ver_merges_heights = ss[s_index] + np.minimum(ks[k_index], ver_hc._merges_heights)
-            ver_merges_heights = ss[s_index] + ver_hc._merges_heights[ver_hc._merges_heights < ks[k_index]]
-            ## same merges in same order
-            #ver_merges = ver_hc._merges
[…]
