+++ persistable/borrowed/dense_mst.pyx
@@ -67,7 +67,9 @@ def stepwise_dendrogram_with_core_distances(int n, const double[:,:] dists, cons
     order = np.argsort(Z_arr[:, 2], kind='mergesort')
     Z_arr = Z_arr[order]
 
+    # do not relabel since we don't really need it, and this makes things easier
+    # when combining different hierarchical clusterings
     # Find correct cluster labels and compute cluster sizes inplace.
-    label(Z_arr, n, n)
+    #label(Z_arr, n, n)
 
     return Z_arr
