+++ persistable/borrowed/prim_mst.pyx
@@ -141,5 +141,7 @@ cpdef np.ndarray[np.double_t, ndim=2] mst_linkage_core_vector(
     
     order = np.argsort(result_arr[:,2], kind='mergesort')
     result_arr = result_arr[order]
-    label(result_arr, dim, dim)
+    # do not relabel since we don't really need it, and this makes things easier
+    # when combining different hierarchical clusterings
+    #label(result_arr, dim, dim)
     return result_arr
