### 2022-12-26 84e0f0c
do not do connection radius automatically at initialization and simplify name of extending clustering
