+++ persistable/persistable.py
@@ -114,17 +114,14 @@ class Persistable:
         self._maxk = self._n_neighbors / X.shape[0]
         # fit the mpspace
         self._mpspace.fit(self._n_neighbors)
-        default_percentile = 0.95
-        # compute and keep robust connection radius
-        self._connection_radius = self._mpspace.connection_radius(default_percentile)
 
     def quick_cluster(
         self,
         n_neighbors: int = 30,
         n_clusters_range=[3, 15],
-        extend_clustering_by_hill_climbing=False,
-        n_iterations_extend_cluster=10,
-        n_neighbors_extend_cluster=5,
+        propagate_labels=False,
+        n_iterations_propagate_labels=30,
+        n_neighbors_propagate_labels=5,
     ):
         """Clusters the dataset with which the Persistable instance was initialized.
 
@@ -141,15 +138,15 @@ class Persistable:
             range of possible numbers of clusters to consider when finding the
             optimum number of clusters.
 
-        extend_clustering_by_hill_climbing: bool, optional, default is False
+        propagate_labels: bool, optional, default is False
             Boolean representing whether or not to extend the clustering to
-            noise points by hill climbing.
+            noise points by propagating labels.
 
-        n_iterations_extend_cluster: int, optional, default is 10
-            How many iterations of hill climbing to perform. More iterations
+        n_iterations_propagate_labels: int, optional, default is 30
+            Maximum number of iterations of label propagation to perform. More iterations
             will cluster more noise points.
 
-        n_neighbors_extend_cluster: int, optional, default is 5
+        n_neighbors_propagate_labels: int, optional, default is 5
             How many neighbors to use in the hill climbing procedure.
 
         returns:
@@ -160,7 +157,9 @@ class Persistable:
 
         """
         k = n_neighbors / self._mpspace._size
-        s = self._connection_radius * 2
+        default_percentile = 0.95
+        s = self._mpspace.connection_radius(default_percentile) * 2
+
         hc = self._mpspace.lambda_linkage([0, k], [s, 0])
         pd = hc.persistence_diagram()
         if pd.shape[0] == 0:
@@ -181,9 +180,9 @@ class Persistable:
             num_clust,
             [0, k],
             [s, 0],
-            extend_clustering_by_hill_climbing=extend_clustering_by_hill_climbing,
-            n_iterations_extend_cluster=n_iterations_extend_cluster,
-            n_neighbors_extend_cluster=n_neighbors_extend_cluster,
+            propagate_labels=propagate_labels,
+            n_iterations_propagate_labels=n_iterations_propagate_labels,
+            n_neighbors_propagate_labels=n_neighbors_propagate_labels,
         )
 
     def cluster(
@@ -191,9 +190,9 @@ class Persistable:
         n_clusters,
         start,
         end,
-        extend_clustering_by_hill_climbing=False,
-        n_iterations_extend_cluster=10,
-        n_neighbors_extend_cluster=5,
+        propagate_labels=False,
+        n_iterations_propagate_labels=30,
+        n_neighbors_propagate_labels=5,
     ):
         """Clusters the dataset with which the Persistable instance was initialized.
 
@@ -214,15 +213,15 @@ class Persistable:
             two-parameter hierarchical clustering used to do persistence-based
             clustering.
 
-        extend_clustering_by_hill_climbing: bool, optional, default is False
+        propagate_labels: bool, optional, default is False
             Boolean representing whether or not to extend the clustering to
-            noise points by hill climbing.
+            noise points by propagating labels.
 
-        n_iterations_extend_cluster: int, optional, default is 10
-            How many iterations of hill climbing to perform. More iterations
+        n_iterations_propagate_labels: int, optional, default is 30
[…]
