### 2022-12-26 125e5bf
stop label propagation if converged
