### 2022-12-27 57e743e
simplify logic in Hilbert function computation
