+++ persistable/persistable.py
@@ -256,7 +256,7 @@ class Persistable:
             threshold = (spers[-n_clusters] + spers[-(n_clusters + 1)]) / 2
         cl = hc.persistence_based_flattening(threshold)
         if propagate_labels:
-            cl = self._mpspace.propagate_labels(
+            cl = self._mpspace._propagate_labels(
                 cl, n_iterations_propagate_labels, n_neighbors_propagate_labels
             )
         return cl
@@ -268,7 +268,7 @@ class Persistable:
         else:
             return self._mpspace.find_end()
 
-    def _compute_vineyard(self, start_end1, start_end2, n_parameters=50, n_jobs=4):
+    def _compute_vineyard(self, start_end1, start_end2, n_parameters, n_jobs=4):
         start1, end1 = start_end1
         start2, end2 = start_end2
         if (
@@ -324,12 +324,8 @@ class Persistable:
 
         # how many more ss than ks (note that getting more ss is very cheap)
         more_s_than_k = 1
-        ss = np.linspace(
-            min_s,
-            max_s + (max_s - min_s) / (granularity * more_s_than_k),
-            granularity * more_s_than_k + 1,
-        )
-        ks = np.linspace(min_k, max_k + (max_k - min_k) / granularity, granularity + 1)
+        ss = np.linspace(min_s, max_s, granularity * more_s_than_k)
+        ks = np.linspace(min_k, max_k, granularity)
         hf = self._mpspace.hilbert_function(ks, ss, n_jobs=n_jobs)
         return ss, ks, hf
 
@@ -697,6 +693,94 @@ class _MetricProbabilitySpace:
                     delayed(run_in_parallel)(startend) for startend in startends
                 )
 
+    def rank_invariant(self, ks, ss, n_jobs):
+        n_s = len(ss)
+        n_k = len(ks)
+        #ks = ks[::-1]
+        startends_horizontal = [ [[ss[0], k], [ss[-1], k]] for k in ks ]
+        startends_vertical = [ [[s, ks[-1]], [s, ks[0]]] for s in ss ]
+        startends = startends_horizontal + startends_vertical
+        hcs = self._lambda_linkage_parallel(startends, n_jobs=n_jobs)
+        hcs_horizontal = hcs[:n_k]
+        hcs_vertical = hcs[n_k:]
+
+        def _splice_hcs(s_index,k_index):
+            # the horizontal hierarchical clustering
+            hor_hc = hcs_horizontal[s_index]
+            # keep only things that happened before or at ss[s_index]
+            hor_heights = hor_hc._heights[hor_hc._heights <= ss[s_index]]
+            hor_merges = hor_hc._merges[hor_hc._merges_heights <= ss[s_index]]
+            hor_merges_heights = hor_hc._merges_heights[hor_hc._merges_heights <= ss[s_index]]
+
+            hor_end = ss[s_index]
+            hor_start = hor_hc._start
+
+            # the vertical hierarchical clustering
+            ver_hc = hcs_vertical[k_index]
+            # push all things that happened before ks[-1] - ks[k_index] there, and index starting from ss[s_index]
+            ver_heights = ss[s_index] + np.maximum(ks[-1] - ks[k_index], ver_hc._heights)
+            ver_merges_heights = ss[s_index] + np.maximum(ks[-1] - ks[k_index], ver_hc._merges_heights)
+
+            # same merges in same order
+            ver_merges = ver_hc._merges
+
+            ver_start = ss[s_index]
+            ver_end = ss[s_index] + ver_hc._end
+
+            heights = np.concatenate((hor_heights,ver_heights))
[…]
