+++ persistable/test/test_persistable.py
@@ -141,8 +141,8 @@ class TestMetricProbabilitySpace(unittest.TestCase):
         p = Persistable(X)
         mps = p._mpspace
 
-        ss = [0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4]
-        ks = [0, 1 / 4 + 0.01, 1 / 2, 3 / 4, 1, 1.1]
+        ss = [0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5]
+        ks = [0, 1 / 4 + 0.01, 1 / 2, 3 / 4, 1]
 
         res = np.array(
             [
@@ -304,18 +304,9 @@ class TestPersistable(unittest.TestCase):
         X = np.array([[0, 0], [1, 0], [1, 1], [3, 0]])
         p = Persistable(X)
 
-        res_ss = [0.0, 0.5625, 1.125, 1.6875, 2.25, 2.8125, 3.375, 3.9375, 4.5]
-        res_ks = [
-            0.0,
-            0.140625,
-            0.28125,
-            0.421875,
-            0.5625,
-            0.703125,
-            0.84375,
-            0.984375,
-            1.125,
-        ]
+        #res_ss = [0.0, 0.5625, 1.125, 1.6875, 2.25, 2.8125, 3.375, 3.9375, 4.5]
+        res_ss = [0.0, 0.5714286, 1.1428571, 1.7142857, 2.2857143, 2.8571429, 3.4285714, 4.]
+        res_ks = [ 0.0, 0.1428571, 0.2857143, 0.4285714, 0.5714286, 0.7142857, 0.8571429, 1 ]
 
         res = np.array(
             [
