### 2022-12-27 fca1e74
fix bug in density estimate computation
