+++ persistable/persistable.py
@@ -272,10 +272,10 @@ class Persistable:
         start1, end1 = start_end1
         start2, end2 = start_end2
         if (
-            start1[0] >= end1[0]
-            or start1[1] <= end1[1]
-            or start2[0] >= end2[0]
-            or start2[1] <= end2[1]
+            start1[0] > end1[0]
+            or start1[1] < end1[1]
+            or start2[0] > end2[0]
+            or start2[1] < end2[1]
         ):
             raise ValueError(
                 "Parameters chosen for vineyard will result in non-monotonic lines!"
@@ -521,7 +521,7 @@ class _MetricProbabilitySpace:
                 neighbor_idx -= 1
                 if self._nn_distance[p,neighbor_idx] < max_density:
                     out_of_range = True
-            density_estimates.append(self._kernel_estimate[p,neighbor_idx])
+            density_estimates.append(self._kernel_estimate[p,neighbor_idx-1])
         if out_of_range:
             warnings.warn(
                 "Don't have enough neighbors to properly compute core scale."
@@ -552,7 +552,9 @@ class _MetricProbabilitySpace:
                     edges.append([i,j])
                     entries.append(min(k_births[i],k_births[j]))
         matrix_entries = np.array(entries)
-        matrix_entries = k_start - np.minimum(k_start, matrix_entries)
+        print("original matrix entries ", matrix_entries)
+        matrix_entries = k_start - np.maximum(k_end, np.minimum(k_start, matrix_entries))
+        print("matrix entries ", matrix_entries)
         edges = np.array(edges,dtype=int)
         graph = csr_matrix( (matrix_entries, (edges[:,0], edges[:,1])),  (self._size, self._size))
 
@@ -570,12 +572,15 @@ class _MetricProbabilitySpace:
         merges_heights = vals
         hc_start = 0
         hc_end = k_start - k_end
-        merges_heights = np.minimum(hc_end, merges_heights)
-        merges_heights = np.maximum(hc_start, merges_heights)
-        core_distances = k_start - np.minimum(k_start, k_births)
+        #merges_heights = np.minimum(hc_end, merges_heights)
+        #merges_heights = np.maximum(hc_start, merges_heights)
+        core_scales = k_start - np.maximum(k_end, np.minimum(k_start, k_births))
+        print("original k births ", k_births)
+        print("births ", core_scales)
+
 
         return _HierarchicalClustering(
-            core_distances, merges, merges_heights, hc_start, hc_end
+            core_scales, merges, merges_heights, hc_start, hc_end
         )
 
     def _lambda_linkage_skew(self, start, end):
@@ -696,36 +701,50 @@ class _MetricProbabilitySpace:
     def rank_invariant(self, ks, ss, n_jobs):
         n_s = len(ss)
         n_k = len(ks)
-        #ks = ks[::-1]
         startends_horizontal = [ [[ss[0], k], [ss[-1], k]] for k in ks ]
         startends_vertical = [ [[s, ks[-1]], [s, ks[0]]] for s in ss ]
         startends = startends_horizontal + startends_vertical
         hcs = self._lambda_linkage_parallel(startends, n_jobs=n_jobs)
         hcs_horizontal = hcs[:n_k]
+        for hc in hcs_horizontal:
+            hc.snap_to_grid(ss)
         hcs_vertical = hcs[n_k:]
+        ks = np.array(ks)
+        for hc in hcs_vertical:
+            hc.snap_to_grid(ks[-1] - ks)
 
         def _splice_hcs(s_index,k_index):
             # the horizontal hierarchical clustering
-            hor_hc = hcs_horizontal[s_index]
-            # keep only things that happened before or at ss[s_index]
-            hor_heights = hor_hc._heights[hor_hc._heights <= ss[s_index]]
[…]
