### 2022-12-27 8813b60
fix bug in vertical slice having to do with edges that appear right at the beginning
