### 2022-12-28 4408188
add first version of computation of rank invariant
