+++ persistable/persistable.py
@@ -534,6 +534,8 @@ class _MetricProbabilitySpace:
 
         indices = np.arange(self._size)
         k_births = self._density_estimate(indices, s_intercept, max_density=k_start)
+        # must add 1 otherwise the edges (below) are treated as not there by mst routine
+        k_births = k_start - np.maximum(k_end, np.minimum(k_start, k_births)) + 1
 
         # metric tree case
         if self._metric in KDTree.valid_metrics + BallTree.valid_metrics:
@@ -550,13 +552,11 @@ class _MetricProbabilitySpace:
             for j in s_neighbors[i]:
                 if j>i:
                     edges.append([i,j])
-                    entries.append(min(k_births[i],k_births[j]))
+                    entries.append(max(k_births[i],k_births[j]))
         matrix_entries = np.array(entries)
-        print("original matrix entries ", matrix_entries)
-        print("matrix pairs", edges)
-        # must add 1 otherwise edge is treated as not there by mst routine!
-        matrix_entries = k_start - np.maximum(k_end, np.minimum(k_start, matrix_entries)) + 1
-        print("matrix entries ", matrix_entries)
+        #print("original matrix entries ", matrix_entries)
+        #print("matrix pairs", edges)
+        #print("matrix entries ", matrix_entries)
         edges = np.array(edges,dtype=int)
         graph = csr_matrix( (matrix_entries, (edges[:,0], edges[:,1])),  (self._size, self._size))
 
@@ -569,21 +569,19 @@ class _MetricProbabilitySpace:
         Js = Js[sort_indices]
         vals = vals[sort_indices]
         merges = np.zeros((vals.shape[0], 2), dtype = int)
-        #label(merges,self._size, vals.shape[0])
         merges[:,0] = Is
         merges[:,1] = Js
         merges_heights = vals
         hc_start = 0
         hc_end = k_start - k_end
-        #merges_heights = np.minimum(hc_end, merges_heights)
-        #merges_heights = np.maximum(hc_start, merges_heights)
-        core_scales = k_start - np.maximum(k_end, np.minimum(k_start, k_births))
-        print("original k births ", k_births)
-        print("births ", core_scales)
+        # we now undo the adding 1
+        core_scales = k_births - 1
+        #print("original k births ", k_births)
+        #print("births ", core_scales)
 
-        print("core scales", core_scales)
-        print("merges ", merges)
-        print("merges_heights ", merges_heights)
+        #print("core scales", core_scales)
+        #print("merges ", merges)
+        #print("merges_heights ", merges_heights)
 
 
         return _HierarchicalClustering(
@@ -724,7 +722,8 @@ class _MetricProbabilitySpace:
             # the horizontal hierarchical clustering
             hor_hc = hcs_horizontal[k_index]
             # keep only things that happened before s_index
-            hor_heights = hor_hc._heights[hor_hc._heights < s_index]
+            hor_heights = hor_hc._heights.copy()
+            hor_heights[hor_heights >= s_index] = s_index + len(ks) - k_index + 1
             hor_merges = hor_hc._merges[hor_hc._merges_heights < s_index]
             hor_merges_heights = hor_hc._merges_heights[hor_hc._merges_heights < s_index]
             hor_end = s_index - 1
@@ -748,22 +747,21 @@ class _MetricProbabilitySpace:
             print("ver heights ", ver_heights)
             print("ver merges heights ", ver_merges_heights)
 
-
-
             ver_start = s_index
             ver_end = s_index + ver_hc._end - k_index
 
-            heights = np.concatenate((hor_heights,ver_heights))
+            heights = np.minimum(hor_heights,ver_heights)
             merges = np.concatenate((hor_merges,ver_merges))
[…]
