+++ persistable/test/test_persistable.py
@@ -156,19 +156,89 @@ class TestMetricProbabilitySpace(unittest.TestCase):
         np.testing.assert_almost_equal(mps.hilbert_function(ks, ss, n_jobs=4), res)
 
     def test_vertical_slice(self):
-        res = np.array([0.72727273, 0.72727273, 0.72727273, 0.72727273, 0.8181818, 0.8181818,0.8181818,0.8181818,0.8181818,0.8181818])
-
-        X = np.array([ [0,0], [1,1], [1,0], [1,-1], [2,0], [3,0], [4,0], [5,1], [5,0], [5,-1], [6,0] ])
+        X = np.array(
+            [
+                [0, 0],
+                [1, 1],
+                [1, 0],
+                [1, -1],
+                [2, 0],
+                [3, 0],
+                [4, 0],
+                [5, 1],
+                [5, 0],
+                [5, -1],
+                [6, 0],
+            ]
+        )
         p = Persistable(X, debug=True)
         mps = p._mpspace
         hc = mps._lambda_linkage_vertical(1, 1, 0)
-        np.testing.assert_almost_equal(res,hc._merges_heights)
+        res = np.array(
+            [
+                0.72727273,
+                0.72727273,
+                0.72727273,
+                0.72727273,
+                0.8181818,
+                0.8181818,
+                0.8181818,
+                0.8181818,
+                0.8181818,
+                0.8181818,
+            ]
+        )
 
-        dist_mat = distance_matrix(X,X)
+
+        np.testing.assert_almost_equal(res, hc._merges_heights)
+
+        dist_mat = distance_matrix(X, X)
         p = Persistable(dist_mat, debug=True, metric="precomputed")
         mps = p._mpspace
         hc = mps._lambda_linkage_vertical(1, 1, 0)
-        np.testing.assert_almost_equal(res,hc._merges_heights)
+        np.testing.assert_almost_equal(res, hc._merges_heights)
+
+        np.testing.assert_almost_equal(
+            np.array([[0.0, 0.1010101], [0.0, 0.1010101]]),
+            mps.lambda_linkage(
+                [1.777777777777778, 0.393939393939394],
+                [1.777777777777778, 0.29292929292929293],
+            ).persistence_diagram(),
+        )
+        np.testing.assert_almost_equal(
+            np.array([[0.0, 0.1], [0.0, 0.1]]),
+            mps.lambda_linkage([1.3, 0.4], [1.3, 0.3]).persistence_diagram(),
+        )
+        np.testing.assert_almost_equal(
+            np.array([[0.0, 0.1], [0.0, 0.1]]),
+            mps.lambda_linkage([1.7, 0.4], [1.7, 0.3]).persistence_diagram(),
+        )
+
+    def test_rank_invariant(self):
+        X = np.array(
+            [
+                [0, 0],
+                [1, 1],
+                [1, 0],
+                [1, -1],
+                [2, 0],
+                [3, 0],
+                [4, 0],
+                [5, 1],
+                [5, 0],
+                [5, -1],
+                [6, 0],
+            ]
+        )
+        p = Persistable(X, debug=True)
+        ks = [0.3, 0.4]
+        ss = [1.3, 1.5]
+
+        ri = p._mpspace.rank_invariant(ks, ss, 2)
+        np.testing.assert_almost_equal(
+            ri,
+            [[[[2, 2], [2, 2]], [[0, 2], [0, 2]]],
+             [[[0, 0], [2, 2]], [[0, 0], [0, 2]]]])
 
 
 
[…]
