### 2022-12-29 0e2ccfb
more robust version of pd computation
