+++ persistable/persistable.py
@@ -706,17 +706,25 @@ class _MetricProbabilitySpace:
     def rank_invariant(self, ks, ss, n_jobs):
         n_s = len(ss)
         n_k = len(ks)
+        ks = np.array(ks)[::-1]
         startends_horizontal = [ [[ss[0], k], [ss[-1], k]] for k in ks ]
-        startends_vertical = [ [[s, ks[-1]], [s, ks[0]]] for s in ss ]
+        startends_vertical = [ [[s, ks[0]], [s, ks[-1]]] for s in ss ]
         startends = startends_horizontal + startends_vertical
         hcs = self._lambda_linkage_parallel(startends, n_jobs=n_jobs)
         hcs_horizontal = hcs[:n_k]
+        i = 0
         for hc in hcs_horizontal:
+            print("snapping horizontal slice at ", ks[i], "births before snapping ", hc._heights)
+            i+=1
             hc.snap_to_grid(ss)
+            print("births after snapping to ", ss, " are: ", hc._heights)
         hcs_vertical = hcs[n_k:]
-        ks = np.array(ks)
+        i = 0
         for hc in hcs_vertical:
-            hc.snap_to_grid(ks[-1] - ks)
+            print("snapping vertical slice at ", ss[i], "births before snapping ", hc._heights)
+            i+=1
+            hc.snap_to_grid(ks[0] - ks[::-1])
+            print("births after snapping to ", ks[0] - ks[::-1], " are: ", hc._heights)
 
         def _splice_hcs(s_index,k_index):
             # the horizontal hierarchical clustering
@@ -748,7 +756,7 @@ class _MetricProbabilitySpace:
             print("ver merges heights ", ver_merges_heights)
 
             ver_start = s_index
-            ver_end = s_index + ver_hc._end - k_index
+            ver_end = s_index + ver_hc._end - k_index + 1
 
             heights = np.minimum(hor_heights,ver_heights)
             merges = np.concatenate((hor_merges,ver_merges))
@@ -778,33 +786,27 @@ class _MetricProbabilitySpace:
                         print("enter bar 2")
                         for i in range(b,s_index+1):
                             print("enter bar 3")
-                            for j in range(s_index,d):
+                            #for j in range(s_index,min(d,s_index + n_k - k_index)):
+                            for j in range(s_index,min(d,s_index + n_k - k_index)):
                                 print("enter bar 4")
+                                print("i ", i, "k_index", k_index, "s_index", s_index, "last", j-s_index+k_index)
                                 ri[i, k_index, s_index, j-s_index+k_index] += 1
 
-
-                ##print(i,j,pd)
-                #if pd.shape[0] > 0:
-                #    for i_ in range(i+1):
-                #        for j_ in range(j,n_k):
-                #            bars_passing = (pd[:,0] <= ss[i_]) & (pd[:,1] >= ss[i] + (ks[-1] - ks[j_]))
-                #            rk = np.count_nonzero(bars_passing)
-                #            ri[i,j,i_,j_] = rk
-        
-        #print(ri[])
-        #0.04, 0.04
-        #0.05, 0.03
-        
         return ri
  
+    ## TODO: use numpy vectorization to speed this up
+    #def kernel_invariant(self, ks, ss, n_jobs):
+    #    ri = self.rank_invariant(ks, ss, n_jobs)
 
-        #for i, pd in enumerate(pds):
-        #    for bar in pd:
-        #        b, d = bar
-        #        start = np.searchsorted(ss[:-1], b)
-        #        end = np.searchsorted(ss[:-1], d)
-        #        hf[i, start:end] += 1
-        #return hf
[…]
