### 2022-12-29 fab446f
fix another bug in rank invariant and add test case
