### 2022-12-29 b1c0d29
add one more test case to rank invariant
