### 2022-12-29 a2e2175
simplify a bunch of conventions
