+++ persistable/persistable.py
@@ -10,7 +10,7 @@ from .borrowed.prim_mst import mst_linkage_core_vector
 from .borrowed.dense_mst import stepwise_dendrogram_with_core_distances
 from .borrowed.dist_metrics import DistanceMetric
 from .auxiliary import lazy_intersection
-from .borrowed.relabel_dendrogram import label
+from .signed_betti_numbers import signed_betti
 import numpy as np
 import warnings
 from sklearn.neighbors import KDTree, BallTree
@@ -298,11 +298,11 @@ class Persistable:
 
     def _compute_hilbert_function(
         self,
-        min_k,
-        max_k,
         min_s,
         max_s,
-        granularity=50,
+        max_k,
+        min_k,
+        granularity,
         n_jobs=4,
     ):
         if min_k >= max_k:
@@ -322,12 +322,43 @@ class Persistable:
                 "max density threshold too large, using " + str(min_k) + " instead."
             )
 
-        # how many more ss than ks (note that getting more ss is very cheap)
-        more_s_than_k = 1
-        ss = np.linspace(min_s, max_s, granularity * more_s_than_k)
-        ks = np.linspace(min_k, max_k, granularity)
-        hf = self._mpspace.hilbert_function(ks, ss, n_jobs=n_jobs)
-        return ss, ks, hf
+        ss = np.linspace(min_s, max_s, granularity)
+        ks = np.linspace(min_k, max_k, granularity)[::-1]
+        hf = self._mpspace.hilbert_function(ss, ks, n_jobs=n_jobs)
+        # the hilbert function has as x axis the ks and as y axis the ss.
+        # the ks are indexed covariantly
+        return ss, ks, hf, signed_betti(hf)
+
+    def _compute_rank_invariant(
+        self,
+        min_s,
+        max_s,
+        max_k,
+        min_k,
+        granularity,
+        n_jobs=4,
+    ):
+        if min_k >= max_k:
+            raise ValueError("min_k must be smaller than max_k.")
+        if min_s >= max_s:
+            raise ValueError("min_s must be smaller than max_s.")
+        if max_k > self._maxk:
+            max_k = min(max_k, self._maxk)
+            warnings.warn(
+                "Not enough neighbors to compute chosen max density threshold, using "
+                + str(max_k)
+                + " instead. If needed, re-initialize the Persistable instance with a larger n_neighbors."
+            )
+        if min_k >= max_k:
+            min_k = max_k / 2
+            warnings.warn(
+                "max density threshold too large, using " + str(min_k) + " instead."
+            )
+
+        ss = np.linspace(min_s, max_s, granularity)
+        ks = np.linspace(min_k, max_k, granularity)[::-1]
+        ri = self._mpspace.rank_invariant(ss, ks, n_jobs=n_jobs)
+        return ss, ks, ri, np.flip(signed_betti(np.flip(ri,(2,3))),(2,3))
 
 
 class _MetricProbabilitySpace:
@@ -509,7 +540,6 @@ class _MetricProbabilitySpace:
                 pd = pers_diag(lower_bound)
                 return [np.max(pd[pd[:, 1] != np.infty][:, 1]), current_k]
 
-
     # given a list of point indices and a radius, return the (unnormalized)
     # kernel density estimate at those points and at that radius
     def _density_estimate(self, point_index, radius, max_density = 1):
@@ -558,20 +588,24 @@ class _MetricProbabilitySpace:
         #print("matrix pairs", edges)
         #print("matrix entries ", matrix_entries)
         edges = np.array(edges,dtype=int)
-        graph = csr_matrix( (matrix_entries, (edges[:,0], edges[:,1])),  (self._size, self._size))
[…]
