+++ persistable/persistable_interactive.py
@@ -108,6 +108,7 @@ N_INTERVALS = "n_intervals"
 IN = "input"
 ST = "state"
 
+
 def empty_figure():
     fig = go.Figure(
         layout=go.Layout(
@@ -1056,7 +1057,7 @@ class PersistableInteractive:
             )
 
             fig.add_trace(
-               go.Heatmap(
+                go.Heatmap(
                     transpose=True,
                     z=ccf,
                     x=x_ticks,
@@ -1066,7 +1067,7 @@ class PersistableInteractive:
                     zmax=max_components,
                     showscale=False,
                     name="",
-               )
+                )
             )
 
             fig.update_traces(colorscale="greys")
@@ -1114,53 +1115,138 @@ class PersistableInteractive:
 
             x_ticks = json.loads(d[STORED_X_TICKS + DATA])
             y_ticks = json.loads(d[STORED_Y_TICKS + DATA])
+            delta_x = (x_ticks[1] - x_ticks[0]) / 2
+            delta_y = (y_ticks[1] - y_ticks[0]) / 2
+
+            max_components = d[INPUT_MAX_COMPONENTS + VALUE]
 
             if False:
                 # need to reset the limits
                 fig.update_yaxes(autorange="reversed")
 
+            def _rgba(color,opacity):
+                if color == "red":
+                    red, green, blue = "255", "0", "0"
+                if color == "green":
+                    red, green, blue = "0", "255", "0"
+                if color == "blue":
+                    red, green, blue = "0", "0", "255"
+                return "rgba(" + red + "," + green + "," + blue + "," + str(opacity) + ")"
+
+            def _draw_bar(xs, ys, color, width=6):
+                return go.Scatter(
+                    x=xs,
+                    y=ys,
+                    marker=dict(color=color),
+                    hoverinfo="skip",
+                    showlegend=False,
+                    mode="lines",
+                    line=dict(width=width),
+                )
+
+            # draw signed barcode
+            if True:
+                sb = np.array(json.loads(d[STORED_SIGNED_BARCODE + DATA]))
+                lx = len(x_ticks)
+                ly = len(y_ticks)
+                traces = []
+                total_width = np.sqrt(lx**2 + ly**2)
+                for i in range(lx):
+                    for j in range(ly):
+                        for i_ in range(i, lx):
+                            for j_ in range(j, ly):
+                                x_coords = np.array(
+                                    [x_ticks[i] - delta_x, x_ticks[i_] - delta_x]
+                                )
+                                y_coords = np.array(
+                                    [y_ticks[j] - delta_y, y_ticks[j_] - delta_y]
+                                )
+                                width = 10 * (
+                                    np.sqrt((i_ - i + 1) ** 2 + (j_ - j + 1) ** 2)
+                                    / total_width
+                                )
+                                min_opacity = 0.3
+                                opacity = min_opacity + (
+                                    np.minimum(np.abs(sb[i, j, i_, j_]), max_components)
+                                    / max_components
+                                ) * (1 - min_opacity)
+                                if sb[i, j, i_, j_] < 0:
+                                    #print(i,j,i_,j_,sb[i, j, i_, j_])
+                                    color = _rgba("red", opacity)
+                                    traces.append(
+                                        _draw_bar(
+                                            x_coords, y_coords, color, width
+                                        )
+                                    )
+                                if sb[i, j, i_, j_] > 0:
+                                    #print(i,j,i_,j_,sb[i, j, i_, j_])
+                                    color = _rgba("blue", opacity)
+                                    traces.append(
+                                        _draw_bar(
[…]
