### 2022-12-30 ba2b46c
speeding up rank invariant computation
