+++ persistable/persistable.py
@@ -9,6 +9,8 @@ from .borrowed._hdbscan_boruvka import (
 from .borrowed.prim_mst import mst_linkage_core_vector
 from .borrowed.dense_mst import stepwise_dendrogram_with_core_distances
 from .borrowed.dist_metrics import DistanceMetric
+from .borrowed.relabel_dendrogram import LinkageUnionFind
+from .borrowed._hdbscan_boruvka import BoruvkaUnionFind
 from .auxiliary import lazy_intersection
 from .signed_betti_numbers import signed_betti, rank_decomposition_hooks_2d, rank_decomposition_rectangles_2d
 import numpy as np
@@ -544,11 +546,9 @@ class _MetricProbabilitySpace:
         density_estimates = []
         out_of_range = False
         for p in point_index:
+            if self._kernel_estimate[p,-1] < max_density:
+                out_of_range = True
             neighbor_idx = np.searchsorted(self._nn_distance[p], radius, side="right")
-            if neighbor_idx == self._nn_distance[p].shape[0]:
-                neighbor_idx -= 1
-                if self._nn_distance[p,neighbor_idx] < max_density:
-                    out_of_range = True
             density_estimates.append(self._kernel_estimate[p,neighbor_idx-1])
         if out_of_range:
             warnings.warn(
@@ -695,10 +695,10 @@ class _MetricProbabilitySpace:
         else:
             return self._lambda_linkage_skew(start,end)
 
-    def lambda_linkage_vineyard(self, startends, n_jobs, tol=_TOL):
+    def lambda_linkage_vineyard(self, startends, n_jobs, reduced=False, tol=_TOL):
         run_in_parallel = lambda startend: self.lambda_linkage(
             startend[0], startend[1]
-        ).persistence_diagram(tol=tol, reduced=True)
+        ).persistence_diagram(tol=tol, reduced=reduced)
         if n_jobs == 1:
             return [run_in_parallel(startend) for startend in startends]
         else:
@@ -735,7 +735,7 @@ class _MetricProbabilitySpace:
                     delayed(run_in_parallel)(startend) for startend in startends
                 )
 
-    def rank_invariant(self, ss, ks, n_jobs):
+    def rank_invariant(self, ss, ks, n_jobs, reduced = False):
         n_s = len(ss)
         n_k = len(ks)
         ks = np.array(ks)
@@ -814,14 +814,14 @@ class _MetricProbabilitySpace:
 
             return _HierarchicalClustering(heights, merges, merges_heights, start, end)
 
-        ri = np.zeros((n_s+1, n_k+1, n_s+1, n_k+1), dtype=int)
+        ri = np.zeros((n_s, n_k, n_s, n_k), dtype=int)
         for s_index in range(n_s):
             for k_index in range(n_k):
                 #DEBUG print("--------------------------")
                 #DEBUG print("s and k index ", s_index, k_index)
                 hc = _splice_hcs(s_index,k_index)
                 #DEBUG print("start end of hierarchical clustering ", hc._start, hc._end)
-                pd = hc.persistence_diagram(reduced=True)
+                pd = hc.persistence_diagram(reduced=reduced)
                 #DEBUG print("persistence diagram ", pd)
                 for bar in pd:
                     b, d = bar
@@ -835,7 +835,7 @@ class _MetricProbabilitySpace:
                             for j in range(s_index,d):
                                 #DEBUG print("enter bar 4")
                                 #DEBUG print("i ", i, "k_index", k_index, "s_index", s_index, "last", j-s_index+k_index)
-                                ri[i, k_index, s_index+1, j-s_index+k_index+1] += 1
+                                ri[i, k_index, s_index, j-s_index+k_index] += 1
 
[…]
