### 2022-12-30 d3cd5ee
move main persistence diagram computation to cython
