+++ persistable/persistable.py
@@ -12,6 +12,7 @@ from .borrowed.dist_metrics import DistanceMetric
 from .borrowed.relabel_dendrogram import LinkageUnionFind
 from .borrowed._hdbscan_boruvka import BoruvkaUnionFind
 from .auxiliary import lazy_intersection
+from .persistence_diagram_h0 import persistence_diagram_h0
 from .signed_betti_numbers import signed_betti, rank_decomposition_hooks_2d, rank_decomposition_rectangles_2d
 import numpy as np
 import warnings
@@ -825,6 +826,7 @@ class _MetricProbabilitySpace:
                 #DEBUG print("persistence diagram ", pd)
                 for bar in pd:
                     b, d = bar
+                    b, d = int(b), int(d)
                     # this if may be unnecessary 
                     #DEBUG print("enter bar")
                     if b <= s_index and d >= s_index:
@@ -884,11 +886,15 @@ class _MetricProbabilitySpace:
 class _HierarchicalClustering:
     def __init__(self, heights, merges, merges_heights, start, end):
         # assumes heights and merges_heights are between start and end
-        self._merges = merges
-        self._merges_heights = merges_heights
-        self._heights = heights
+        self._merges = np.array(merges, dtype=int)
+        self._merges_heights = np.array(merges_heights, dtype=float)
+        self._heights = np.array(heights, dtype=float)
         self._start = start
         self._end = end
+        if self._merges.shape[0] == 0:
+            self._merges = np.array([[0,0]], dtype=int)
+            self._merges_heights = np.array([self._end], dtype=float)
+
 
     def snap_to_grid(self, grid):
         def _snap_array(grid, arr):
@@ -1026,103 +1032,7 @@ class _HierarchicalClustering:
         return res
 
     def persistence_diagram(self, reduced=False, tol=_TOL):
-        end = self._end
-        heights = self._heights
-        merges = self._merges
-        merges_heights = self._merges_heights
-        n_points = heights.shape[0]
-        n_merges = merges.shape[0]
-        # this orders the point by appearance
-        appearances = np.argsort(heights)
-        # contains the current clusters
-        ##uf = DisjointSet()
-        uf = BoruvkaUnionFind(len(heights))
-        # contains the birth time of clusters that are alive
-        clusters_birth = {}
-        clusters_died = {}
-        # contains the persistence diagram
-        pd = []
-        # height index
-        hind = 0
-        # merge index
-        mind = 0
-        if len(appearances) == 0:
-            return np.array([])
-        current_appearence_height = heights[appearances[0]]
-        if len(merges_heights) == 0:
-            current_merge_height = end
-        else:
-            current_merge_height = merges_heights[0]
-        while True:
-            # while there is no merge
-            while (
-                hind < n_points
-                and heights[appearances[hind]] <= current_merge_height
-                and heights[appearances[hind]] < end
-            ):
-                # add all points that are born as new clusters
-                ##uf.add(appearances[hind])
-                ##clusters_birth[appearances[hind]] = heights[appearances[hind]]
-                clusters_birth[uf.find(appearances[hind])] = heights[appearances[hind]]
-                hind += 1
-                if hind == n_points:
-                    current_appearence_height = end
-                else:
-                    current_appearence_height = heights[appearances[hind]]
-            # while there is no cluster being born
-            while (
-                mind < n_merges
-                and merges_heights[mind] < current_appearence_height
-                and merges_heights[mind] < end
-            ):
-                xy = merges[mind]
-                x, y = xy
-                rx = uf.find(x)
-                ry = uf.find(y)
-                # if they were not already merged
-                if rx != ry:
-                    # if both clusters are alive, merge them and add a bar to the pd
-                    if rx not in clusters_died and ry not in clusters_died:
-                        bx = clusters_birth[rx]
[…]
