### 2022-12-30 386b948
speed up barcode a bit more and reorganize persistable interactive
