+++ persistable/persistable_interactive.py
@@ -52,7 +52,8 @@ STORED_X_TICKS = "stored-x-ticks-"
 STORED_Y_TICKS = "stored-y-ticks-"
 STORED_CCF_DRAWING = "stored-ccf-drawing-"
 STORED_BETTI = "stored-betti-"
-STORED_SIGNED_BARCODE = "stored-signed-barcode-"
+STORED_SIGNED_BARCODE_RECTANGLES = "stored-signed-barcode-rectangles-"
+STORED_SIGNED_BARCODE_HOOKS = "stored-signed-barcode-hooks-"
 MIN_DIST_SCALE = "min-dist-scale-"
 MAX_DIST_SCALE = "max-dist-scale-"
 MIN_DENSITY_THRESHOLD = "min-density-threshold-"
@@ -66,6 +67,7 @@ INPUT_GRANULARITY_CCF = "input-granularity-ccf-"
 INPUT_NUM_JOBS_CCF = "input-num-jobs-ccf-"
 INPUT_MAX_COMPONENTS = "input-max-components-"
 INPUT_SIGNED_BETTI_NUMBERS = "input-signed-betti-numbers-"
+INPUT_Y_COVARIANT = "input-y-covariant-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
 PV_DETAILS = "pv-details-"
@@ -294,7 +296,8 @@ class PersistableInteractive:
                 # contains the signed betti numbers as a list of lists
                 dcc.Store(id=STORED_BETTI),
                 # contains the signed barcode a list of lists of ...
-                dcc.Store(id=STORED_SIGNED_BARCODE),
+                dcc.Store(id=STORED_SIGNED_BARCODE_RECTANGLES),
+                dcc.Store(id=STORED_SIGNED_BARCODE_HOOKS),
                 # contains the vineyard as a vineyard object
                 dcc.Store(id=STORED_PV),
                 # contains the basic prominence vineyard plot as a plotly figure
@@ -430,12 +433,12 @@ class PersistableInteractive:
                                                                 ),
                                                                 html.Span(
                                                                     className="name",
-                                                                    children="Signed Betti numbers",
+                                                                    children="y axis",
                                                                 ),
                                                                 dcc.RadioItems(
-                                                                    ["On", "Off"],
-                                                                    "Off",
-                                                                    id=INPUT_SIGNED_BETTI_NUMBERS,
+                                                                    ["Covariant", "Contravariant"],
+                                                                    "Covariant",
+                                                                    id=INPUT_Y_COVARIANT,
                                                                     className="small-value",
                                                                 ),
                                                             ],
@@ -1100,12 +1103,14 @@ class PersistableInteractive:
                 [FIXED_PARAMETERS, DATA, IN],
                 [STORED_PD, DATA, ST],
                 [INPUT_GAP, VALUE, ST],
-                [INPUT_SIGNED_BETTI_NUMBERS, VALUE, IN],
+                #[INPUT_SIGNED_BETTI_NUMBERS, VALUE, IN],
+                [INPUT_Y_COVARIANT, VALUE, IN],
                 [STORED_BETTI, DATA, ST],
                 [STORED_X_TICKS, DATA, ST],
[…]

[…]
