### 2022-12-30 7fac176
use non-reduced rank invariant in tests
