### 2022-12-30 549d8f7
improve displaying of rank decompositions
