### 2022-12-31 5a79d69
save signed barcode as sparse
