+++ persistable/persistable_interactive.py
@@ -436,7 +436,10 @@ class PersistableInteractive:
                                                                     children="y axis",
                                                                 ),
                                                                 dcc.RadioItems(
-                                                                    ["Covariant", "Contravariant"],
+                                                                    [
+                                                                        "Covariant",
+                                                                        "Contravariant",
+                                                                    ],
                                                                     "Covariant",
                                                                     id=INPUT_Y_COVARIANT,
                                                                     className="small-value",
@@ -1103,7 +1106,7 @@ class PersistableInteractive:
                 [FIXED_PARAMETERS, DATA, IN],
                 [STORED_PD, DATA, ST],
                 [INPUT_GAP, VALUE, ST],
-                #[INPUT_SIGNED_BETTI_NUMBERS, VALUE, IN],
+                # [INPUT_SIGNED_BETTI_NUMBERS, VALUE, IN],
                 [INPUT_Y_COVARIANT, VALUE, IN],
                 [STORED_BETTI, DATA, ST],
                 [STORED_X_TICKS, DATA, ST],
@@ -1122,6 +1125,8 @@ class PersistableInteractive:
             y_ticks = json.loads(d[STORED_Y_TICKS + DATA])
             delta_x = (x_ticks[1] - x_ticks[0]) / 2
             delta_y = (y_ticks[1] - y_ticks[0]) / 2
+            x_ticks.append(np.array(x_ticks[-1]) + 2 * delta_x)
+            y_ticks.append(np.array(y_ticks[-1]) + 2 * delta_y)
 
             max_components = d[INPUT_MAX_COMPONENTS + VALUE]
 
@@ -1129,18 +1134,20 @@ class PersistableInteractive:
                 # need to reset the limits
                 fig.update_yaxes(autorange="reversed")
 
-            def _rgba(color,opacity):
+            def _rgba(color, opacity):
                 if color == "red":
                     red, green, blue = "255", "0", "0"
                 if color == "green":
                     red, green, blue = "0", "255", "0"
                 if color == "blue":
                     red, green, blue = "0", "0", "255"
-                return "rgba(" + red + "," + green + "," + blue + "," + str(opacity) + ")"
+                return (
+                    "rgba(" + red + "," + green + "," + blue + "," + str(opacity) + ")"
+                )
 
             def _draw_bar(xs, ys, color, width=6, endpoints=False, size=5):
                 mode = "markers+lines" if endpoints else "lines"
-                marker_styles = ["arrow-right", "arrow-left"]
+                marker_styles = ["diamond", "diamond"]
                 return go.Scatter(
                     x=xs,
                     y=ys,
@@ -1158,74 +1165,83 @@ class PersistableInteractive:
                 USE_RECTANGLES = True
                 using_rectangles = USE_RECTANGLES
                 if using_rectangles:
-                    sb = np.array(json.loads(d[STORED_SIGNED_BARCODE_RECTANGLES + DATA]))
+                    sb = np.array(
+                        json.loads(d[STORED_SIGNED_BARCODE_RECTANGLES + DATA])
+                    )
                 else:
                     sb = np.array(json.loads(d[STORED_SIGNED_BARCODE_HOOKS + DATA]))
                 lx = len(x_ticks)
                 ly = len(y_ticks)
                 traces = []
                 if using_rectangles:
-                    total_width = min(lx,ly)
+                    total_width = min(lx, ly)
                 else:
                     total_width = np.sqrt(lx**2 + ly**2)
-                for i in range(lx):
-                    for j in range(ly):
-                        for i_ in range(i, lx):
-                            for j_ in range(j, ly):
-                                mult = sb[i, j, i_, j_]
-                                if mult != 0:
-                                    #if using_rectangles:
-                                    #    i_ += 1
-                                    #    j_ += 1
-                                    #else:
[…]
