### 2022-12-31 33ab702
speed up rank invariant computation with parallelization
