+++ persistable/persistable.py
@@ -11,19 +11,41 @@ from .borrowed.dense_mst import stepwise_dendrogram_with_core_distances
 from .borrowed.dist_metrics import DistanceMetric
 from .auxiliary import lazy_intersection
 from .persistence_diagram_h0 import persistence_diagram_h0
-from .signed_betti_numbers import signed_betti, rank_decomposition_2d_rectangles, rank_decomposition_2d_rectangles_to_hooks
+from .signed_betti_numbers import (
+    signed_betti,
+    rank_decomposition_2d_rectangles,
+    rank_decomposition_2d_rectangles_to_hooks,
+)
 import numpy as np
 import warnings
 from sklearn.neighbors import KDTree, BallTree
 from scipy.cluster.hierarchy import DisjointSet
 from scipy.sparse import csr_matrix
-from scipy.sparse.csgraph import minimum_spanning_tree as sparse_matrix_minimum_spanning_tree
+from scipy.sparse.csgraph import (
+    minimum_spanning_tree as sparse_matrix_minimum_spanning_tree,
+)
 from scipy.stats import mode
 from joblib import Parallel, delayed
 from joblib.parallel import cpu_count
 
 
 _TOL = 1e-08
+
+
+def parallel_computation(function, inputs, n_jobs, debug=False, threading=False):
+    if n_jobs == 1:
+        return [function(inp) for inp in inputs]
+    else:
+        verbose = 11 if debug else 0
+        n_jobs = min(cpu_count(), n_jobs)
+        if threading:
+            return Parallel(n_jobs=n_jobs, backend="threading", verbose=verbose)(
+                delayed(function)(inp) for inp in inputs
+            )
+        else:
+            return Parallel(n_jobs=n_jobs, verbose=verbose)(
+                delayed(function)(inp) for inp in inputs
+            )
 
 
 class Persistable:
@@ -390,7 +412,7 @@ class _MetricProbabilitySpace:
         self._kernel_estimate = None
         self._n_neighbors = None
         self._maxs = None
-        #self._tol = _TOL
+        # self._tol = _TOL
         if metric in KDTree.valid_metrics:
             self._tree = KDTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
         elif metric in BallTree.valid_metrics:
@@ -542,18 +564,16 @@ class _MetricProbabilitySpace:
 
     # given a list of point indices and a radius, return the (unnormalized)
     # kernel density estimate at those points and at that radius
-    def _density_estimate(self, point_index, radius, max_density = 1):
+    def _density_estimate(self, point_index, radius, max_density=1):
         density_estimates = []
         out_of_range = False
         for p in point_index:
-            if self._kernel_estimate[p,-1] < max_density:
+            if self._kernel_estimate[p, -1] < max_density:
                 out_of_range = True
             neighbor_idx = np.searchsorted(self._nn_distance[p], radius, side="right")
-            density_estimates.append(self._kernel_estimate[p,neighbor_idx-1])
+            density_estimates.append(self._kernel_estimate[p, neighbor_idx - 1])
         if out_of_range:
-            warnings.warn(
-                "Don't have enough neighbors to properly compute core scale."
-            )
+            warnings.warn("Don't have enough neighbors to properly compute core scale.")
         return np.array(density_estimates)
 
     def _lambda_linkage_vertical(self, s_intercept, k_start, k_end):
@@ -567,39 +587,41 @@ class _MetricProbabilitySpace:
 
         # metric tree case
         if self._metric in KDTree.valid_metrics + BallTree.valid_metrics:
-            s_neighbors = self._tree.query_radius(self._points,s_intercept)
+            s_neighbors = self._tree.query_radius(self._points, s_intercept)
         # dense distance matrix case
         else:
             s_neighbors = []
[…]
