### 2022-12-31 09a98ce
add computation of rank separate from hilbert function
