+++ persistable/persistable_interactive.py
@@ -48,8 +48,10 @@ CCF_PLOT = "ccf-plot-"
 DISPLAY_LINES_SELECTION = "display-lines-selection-"
 ENDPOINT_SELECTION = "endpoint-selection-"
 STORED_CCF = "stored-ccf-"
-STORED_X_TICKS = "stored-x-ticks-"
-STORED_Y_TICKS = "stored-y-ticks-"
+STORED_X_TICKS_CCF = "stored-x-ticks-ccf-"
+STORED_Y_TICKS_CCF = "stored-y-ticks-ccf-"
+STORED_X_TICKS_RI = "stored-x-ticks-ri-"
+STORED_Y_TICKS_RI = "stored-y-ticks-ri-"
 STORED_CCF_DRAWING = "stored-ccf-drawing-"
 STORED_BETTI = "stored-betti-"
 STORED_SIGNED_BARCODE_RECTANGLES = "stored-signed-barcode-rectangles-"
@@ -63,11 +65,18 @@ PARAMETER_SELECTION_DIV = "parameter-selection-div-"
 DISPLAY_PARAMETER_SELECTION = "display-parameter-selection-"
 COMPUTE_CCF_BUTTON = "compute-ccf-button-"
 STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
+COMPUTE_RI_BUTTON = "compute-ri-button-"
+STOP_COMPUTE_RI_BUTTON = "stop-compute-ri-button-"
 INPUT_GRANULARITY_CCF = "input-granularity-ccf-"
+INPUT_GRANULARITY_RI = "input-granularity-ri-"
 INPUT_NUM_JOBS_CCF = "input-num-jobs-ccf-"
+INPUT_NUM_JOBS_RI = "input-num-jobs-ri-"
 INPUT_MAX_COMPONENTS = "input-max-components-"
+INPUT_MAX_RI = "input-max-ri-"
+INPUT_MIN_LENGTH_RI = "input-min-length-bars-ri-"
 INPUT_SIGNED_BETTI_NUMBERS = "input-signed-betti-numbers-"
 INPUT_Y_COVARIANT = "input-y-covariant-"
+INPUT_DISPLAY_RI = "input-display-ri-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
 CCF_EXTRAS = "ccf-extras-"
@@ -92,6 +101,7 @@ STORED_PD = "stored-pd-"
 
 STORED_CCF_COMPUTATION_WARNINGS = "stored-ccf-computation-warnings-"
 STORED_PV_COMPUTATION_WARNINGS = "stored-pv-computation-warnings-"
+STORED_RI_COMPUTATION_WARNINGS = "stored-ri-computation-warnings-"
 # STORED_PD_COMPUTATION_WARNINGS = "stored-pd-computation-warnings-"
 
 EXPORTED_PARAMETER = "exported-parameter-"
@@ -262,6 +272,7 @@ class PersistableInteractive:
         default_max_s = end[0]
         # default_s_step = (default_max_s - default_min_s) / 100
         default_granularity_ccf = 100
+        default_granularity_ri = 20
         default_granularity_pv = 40
         default_num_jobs = 4
         default_max_dim = 15
@@ -285,20 +296,610 @@ class PersistableInteractive:
         default_y_end_second_line = (default_min_k + default_max_k) * (1 / 2)
 
         self._app.title = "Persistable"
+
+        ccf_inputs = (
+            html.Details(
+                id=CCF_DETAILS,
+                children=[
+                    html.Summary("Inputs"),
+                    html.Div(
+                        className="parameters",
+                        children=[
+                            html.Div(
+                                className="parameter-double",
[…]
