### 2022-12-31 6d82fc6
add option to filter bars by length
