### 2022-12-31 a31f7ce
add rank decomposition by hooks and reduced homology
