### 2023-01-01 ec488bc
improve gui for rank decomposition and fix displaying of barcode
