+++ persistable/persistable_interactive.py
@@ -361,6 +361,34 @@ class PersistableInteractive:
                                     ),
                                 ],
                             ),
+                            # html.Div(
+                            #    className="row",
+                            #    children=[
+                            #        html.Div(
+                            #            className="column",
+                            #            children=[
+                            #                html.Div(
+                            #                    className="parameter-single",
+                            #                    children=[
+                            #                        html.Span(
+                            #                            className="name",
+                            #                            children="Granularity",
+                            #                        ),
+                            #                        dcc.Input(
+                            #                            id=INPUT_GRANULARITY_CCF,
+                            #                            className="small-value",
+                            #                            type="number",
+                            #                            value=default_granularity_ccf,
+                            #                            min=min_granularity,
+                            #                            max=max_granularity,
+                            #                            debounce=True,
+                            #                        ),
+                            #                    ],
+                            #                )
+                            #            ],
+                            #        )
+                            #    ],
+                            # ),
                             html.Div(
                                 className="parameter-single",
                                 children=[
@@ -397,7 +425,7 @@ class PersistableInteractive:
                                 children=[
                                     html.Span(
                                         className="name",
-                                        children="Number of cores computation",
+                                        children="# cores computation",
                                     ),
                                     dcc.Input(
                                         className="small-value",
@@ -411,14 +439,14 @@ class PersistableInteractive:
                                     ),
                                     html.Span(
                                         className="name",
-                                        children="y axis",
+                                        children="Y axis",
                                     ),
                                     dcc.RadioItems(
                                         [
-                                            "Covariant",
-                                            "Contravariant",
+                                            "Cov",
+                                            "Contr",
                                         ],
-                                        "Covariant",
+                                        "Cov",
                                         id=INPUT_Y_COVARIANT,
                                         className="small-value",
                                     ),
@@ -512,6 +540,7 @@ class PersistableInteractive:
                         id=CCF_EXTRAS,
                         children=[
                             html.Summary("Extras"),
+                            html.H3("Rank decomposition"),
                             html.Div(
                                 className="parameters",
                                 children=[
@@ -520,7 +549,7 @@ class PersistableInteractive:
                                         children=[
                                             html.Span(
                                                 className="name",
-                                                children="Granularity rank decomposition",
+                                                children="Granularity",
                                             ),
                                             dcc.Input(
                                                 id=INPUT_GRANULARITY_RI,
@@ -564,57 +593,7 @@ class PersistableInteractive:
                                             ),
                                             html.Span(
                                                 className="name",
-                                                children="Decompose by",
-                                            ),
-                                            dcc.RadioItems(
-                                                [
-                                                    "Rectangles",
-                                                    "Hooks",
-                                                ],
-                                                "Rectangles",
-                                                id=INPUT_DECOMPOSE_BY_RI,
-                                                className="small-value",
-                                            ),
-                                        ],
-                                    ),
-                                    html.Div(
-                                        className="parameter-single",
-                                        children=[
-                                            html.Span(
-                                                className="name",
-                                                children="Rank decomposition reduced homology",
-                                            ),
-                                            dcc.RadioItems(
-                                                [
-                                                    "Yes",
[…]
