+++ persistable/persistable_interactive.py
@@ -104,7 +104,6 @@ STORED_PD = "stored-pd-"
 STORED_CCF_COMPUTATION_WARNINGS = "stored-ccf-computation-warnings-"
 STORED_PV_COMPUTATION_WARNINGS = "stored-pv-computation-warnings-"
 STORED_RI_COMPUTATION_WARNINGS = "stored-ri-computation-warnings-"
-# STORED_PD_COMPUTATION_WARNINGS = "stored-pd-computation-warnings-"
 
 EXPORTED_PARAMETER = "exported-parameter-"
 
@@ -201,9 +200,6 @@ class PersistableInteractive:
             self._layout_gui()
             self._register_callbacks(self._persistable, self._debug)
 
-            # import logging
-            # self._app.logger.setLevel(logging.WARNING)
-            # logging.getLogger("werkzeug").setLevel(logging.ERROR)
             self._app.run_server(port=port, mode="inline", debug=debug)
         else:
             self._app = dash.Dash(
@@ -281,6 +277,7 @@ class PersistableInteractive:
         default_max_vines = 15
         min_granularity = 4
         max_granularity = 512
+        max_granularity_ri = 64
         min_granularity_vineyard = 1
         max_granularity_vineyard = max_granularity
         defr = 6
@@ -361,34 +358,6 @@ class PersistableInteractive:
                                     ),
                                 ],
                             ),
-                            # html.Div(
-                            #    className="row",
-                            #    children=[
-                            #        html.Div(
-                            #            className="column",
-                            #            children=[
-                            #                html.Div(
-                            #                    className="parameter-single",
-                            #                    children=[
-                            #                        html.Span(
-                            #                            className="name",
-                            #                            children="Granularity",
-                            #                        ),
-                            #                        dcc.Input(
-                            #                            id=INPUT_GRANULARITY_CCF,
-                            #                            className="small-value",
-                            #                            type="number",
-                            #                            value=default_granularity_ccf,
-                            #                            min=min_granularity,
-                            #                            max=max_granularity,
-                            #                            debounce=True,
-                            #                        ),
-                            #                    ],
-                            #                )
-                            #            ],
-                            #        )
-                            #    ],
-                            # ),
                             html.Div(
                                 className="parameter-single",
                                 children=[
@@ -557,7 +526,7 @@ class PersistableInteractive:
                                                 type="number",
                                                 value=default_granularity_ri,
                                                 min=min_granularity,
-                                                max=max_granularity,
+                                                max=max_granularity_ri,
                                                 debounce=True,
                                             ),
                                             html.Span(
@@ -971,12 +940,10 @@ class PersistableInteractive:
                 dcc.Store(id=STORED_CCF_COMPUTATION_WARNINGS, data=json.dumps(" ")),
                 dcc.Store(id=STORED_RI_COMPUTATION_WARNINGS, data=json.dumps(" ")),
                 dcc.Store(id=STORED_PV_COMPUTATION_WARNINGS, data=json.dumps(" ")),
-                # dcc.Store(id=STORED_PD_COMPUTATION_WARNINGS),
                 #
                 dcc.Store(id=FIXED_PARAMETERS, data=json.dumps([])),
                 #
                 html.Div(id=EXPORTED_PARAMETER, hidden=True),
-                # html.Div(id=EXPORTED_PARAMETER_2, hidden=True),
                 html.Div(
                     className="grid",
                     children=[
@@ -1417,16 +1384,6 @@ class PersistableInteractive:
[…]
