+++ persistable/persistable.py
@@ -414,7 +414,6 @@ class _MetricProbabilitySpace:
         self._kernel_estimate = None
         self._n_neighbors = None
         self._maxs = None
-        # self._tol = _TOL
         if metric in KDTree.valid_metrics:
             self._tree = KDTree(X, metric=metric, leaf_size=leaf_size, **kwargs)
         elif metric in BallTree.valid_metrics:
@@ -604,9 +603,6 @@ class _MetricProbabilitySpace:
                     edges.append([i, j])
                     entries.append(max(k_births[i], k_births[j]))
         matrix_entries = np.array(entries)
-        # print("original matrix entries ", matrix_entries)
-        # print("matrix pairs", edges)
-        # print("matrix entries ", matrix_entries)
         edges = np.array(edges, dtype=int)
         if len(edges) > 0:
             graph = csr_matrix(
@@ -632,12 +628,6 @@ class _MetricProbabilitySpace:
         hc_end = k_start - k_end
         # we now undo the adding 1
         core_scales = k_births - 1
-        # print("original k births ", k_births)
-        # print("births ", core_scales)
-
-        # print("core scales", core_scales)
-        # print("merges ", merges)
-        # print("merges_heights ", merges_heights)
 
         return _HierarchicalClustering(
             core_scales, merges, merges_heights, hc_start, hc_end
@@ -731,42 +721,6 @@ class _MetricProbabilitySpace:
             threading=self._threading,
         )
 
-        ##if n_jobs == 1:
-        ##    return [run_in_parallel(startend) for startend in startends]
-        ##else:
-        ##    verbose = 11 if self._debug else 0
-        ##    n_jobs = min(cpu_count(), n_jobs)
-        ##    # x = Parallel(n_jobs=n_jobs, backend="multiprocessing", verbose=verbose)(
-        ##    #    delayed(self.lambda_linkage)(startend[0], startend[1]) for startend in startends
-        ##    # )
-        ##    # return [hc.persistence_diagram(tol=tol) for hc in x]
-        ##    if self._threading:
-        ##        return Parallel(n_jobs=n_jobs, backend="threading", verbose=verbose)(
-        ##            delayed(run_in_parallel)(startend) for startend in startends
-        ##        )
-        ##    else:
-        ##        return Parallel(n_jobs=n_jobs, verbose=verbose)(
-        ##            delayed(run_in_parallel)(startend) for startend in startends
-        ##        )
-
-    # def _lambda_linkage_parallel(self, startends, n_jobs):
-    #    run_in_parallel = lambda startend: self.lambda_linkage(
-    #        startend[0], startend[1]
-    #    )
-    #    if n_jobs == 1:
-    #        return [run_in_parallel(startend) for startend in startends]
-    #    else:
-    #        verbose = 11 if self._debug else 0
-    #        n_jobs = min(cpu_count(), n_jobs)
-    #        if self._threading:
-    #            return Parallel(n_jobs=n_jobs, backend="threading", verbose=verbose)(
-    #                delayed(run_in_parallel)(startend) for startend in startends
-    #            )
-    #        else:
-    #            return Parallel(n_jobs=n_jobs, verbose=verbose)(
-    #                delayed(run_in_parallel)(startend) for startend in startends
-    #            )
-
     def rank_invariant(self, ss, ks, n_jobs, reduced=False):
         n_s = len(ss)
         n_k = len(ks)
@@ -783,19 +737,11 @@ class _MetricProbabilitySpace:
             threading=self._threading,
         )
         hcs_horizontal = hcs[:n_k]
-        # DEBUG i = 0
         for hc in hcs_horizontal:
[…]
