### 2023-01-03 ac6488e
fix max neighbors bug
