### 2023-01-03 3e82874
add test case checking that k can be 1 for horizontal slice of degree rips
