+++ persistable/persistable_interactive.py
@@ -8,7 +8,8 @@ import plotly.graph_objects as go
 import plotly
 from plotly.colors import sample_colorscale
 import json
-#import diskcache
+
+# import diskcache
 import dash
 from dash import dcc, html, DiskcacheManager, ctx
 from dash.exceptions import PreventUpdate
@@ -47,7 +48,14 @@ CCF_PLOT = "ccf-plot-"
 DISPLAY_LINES_SELECTION = "display-lines-selection-"
 ENDPOINT_SELECTION = "endpoint-selection-"
 STORED_CCF = "stored-ccf-"
+STORED_X_TICKS_CCF = "stored-x-ticks-ccf-"
+STORED_Y_TICKS_CCF = "stored-y-ticks-ccf-"
+STORED_X_TICKS_RI = "stored-x-ticks-ri-"
+STORED_Y_TICKS_RI = "stored-y-ticks-ri-"
 STORED_CCF_DRAWING = "stored-ccf-drawing-"
+STORED_BETTI = "stored-betti-"
+STORED_SIGNED_BARCODE_RECTANGLES = "stored-signed-barcode-rectangles-"
+STORED_SIGNED_BARCODE_HOOKS = "stored-signed-barcode-hooks-"
 MIN_DIST_SCALE = "min-dist-scale-"
 MAX_DIST_SCALE = "max-dist-scale-"
 MIN_DENSITY_THRESHOLD = "min-density-threshold-"
@@ -57,11 +65,23 @@ PARAMETER_SELECTION_DIV = "parameter-selection-div-"
 DISPLAY_PARAMETER_SELECTION = "display-parameter-selection-"
 COMPUTE_CCF_BUTTON = "compute-ccf-button-"
 STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
+COMPUTE_RI_BUTTON = "compute-ri-button-"
+STOP_COMPUTE_RI_BUTTON = "stop-compute-ri-button-"
 INPUT_GRANULARITY_CCF = "input-granularity-ccf-"
+INPUT_GRANULARITY_RI = "input-granularity-ri-"
 INPUT_NUM_JOBS_CCF = "input-num-jobs-ccf-"
+INPUT_NUM_JOBS_RI = "input-num-jobs-ri-"
 INPUT_MAX_COMPONENTS = "input-max-components-"
+INPUT_MAX_RI = "input-max-ri-"
+INPUT_MIN_LENGTH_RI = "input-min-length-bars-ri-"
+INPUT_SIGNED_BETTI_NUMBERS = "input-signed-betti-numbers-"
+INPUT_Y_COVARIANT = "input-y-covariant-"
+INPUT_DISPLAY_RI = "input-display-ri-"
+INPUT_DECOMPOSE_BY_RI = "input-decompose-by-ri-"
+INPUT_REDUCED_HOMOLOGY_RI = "input-reduced-homology-ri-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
+CCF_EXTRAS = "ccf-extras-"
 PV_DETAILS = "pv-details-"
 PV_PLOT_CONTROLS_DIV = "pv-plot-controls-div-"
 LOG = "log-"
@@ -83,7 +103,7 @@ STORED_PD = "stored-pd-"
 
 STORED_CCF_COMPUTATION_WARNINGS = "stored-ccf-computation-warnings-"
 STORED_PV_COMPUTATION_WARNINGS = "stored-pv-computation-warnings-"
-# STORED_PD_COMPUTATION_WARNINGS = "stored-pd-computation-warnings-"
[…]
