### 2023-01-03 214f2ab
add signed decomposition of rank invariant and cythonize persistence diagram computation (#16)
* add betti number computation

* add hierarchical clustering over vertical slice

* do not relabel stepwise dendrogram

* do not do connection radius automatically at initialization and simplify name of extending clustering

* stop label propagation if converged

* fix bug in density estimate computation

* more robust version of pd computation

* simplify conventions for horizontal and vertical coordinates

* move main persistence diagram computation to cython

* add test case checking that k can b
[…]
