+++ persistable/persistable.py
@@ -10,16 +10,42 @@ from .borrowed.prim_mst import mst_linkage_core_vector
 from .borrowed.dense_mst import stepwise_dendrogram_with_core_distances
 from .borrowed.dist_metrics import DistanceMetric
 from .auxiliary import lazy_intersection
+from .persistence_diagram_h0 import persistence_diagram_h0
+from .signed_betti_numbers import (
+    signed_betti,
+    rank_decomposition_2d_rectangles,
+    rank_decomposition_2d_rectangles_to_hooks,
+)
 import numpy as np
 import warnings
 from sklearn.neighbors import KDTree, BallTree
 from scipy.cluster.hierarchy import DisjointSet
+from scipy.sparse import csr_matrix
+from scipy.sparse.csgraph import (
+    minimum_spanning_tree as sparse_matrix_minimum_spanning_tree,
+)
 from scipy.stats import mode
 from joblib import Parallel, delayed
 from joblib.parallel import cpu_count
 
 
 _TOL = 1e-08
+
+
+def parallel_computation(function, inputs, n_jobs, debug=False, threading=False):
+    if n_jobs == 1:
+        return [function(inp) for inp in inputs]
+    else:
+        verbose = 11 if debug else 0
+        n_jobs = min(cpu_count(), n_jobs)
+        if threading:
+            return Parallel(n_jobs=n_jobs, backend="threading", verbose=verbose)(
+                delayed(function)(inp) for inp in inputs
+            )
+        else:
+            return Parallel(n_jobs=n_jobs, verbose=verbose)(
+                delayed(function)(inp) for inp in inputs
+            )
 
 
 class Persistable:
@@ -111,17 +137,14 @@ class Persistable:
         self._maxk = self._n_neighbors / X.shape[0]
         # fit the mpspace
         self._mpspace.fit(self._n_neighbors)
-        default_percentile = 0.95
-        # compute and keep robust connection radius
-        self._connection_radius = self._mpspace.connection_radius(default_percentile)
 
     def quick_cluster(
         self,
         n_neighbors: int = 30,
         n_clusters_range=[3, 15],
-        extend_clustering_by_hill_climbing=False,
-        n_iterations_extend_cluster=10,
-        n_neighbors_extend_cluster=5,
+        propagate_labels=False,
+        n_iterations_propagate_labels=30,
+        n_neighbors_propagate_labels=5,
     ):
         """Find parameters automatically and cluster dataset passed at initialization.
 
@@ -138,15 +161,15 @@ class Persistable:
             range of possible numbers of clusters to consider when finding the
             optimum number of clusters.
 
-        extend_clustering_by_hill_climbing: bool, optional, default is False
+        propagate_labels: bool, optional, default is False
             Boolean representing whether or not to extend the clustering to
-            noise points by hill climbing.
+            noise points by propagating labels.
 
-        n_iterations_extend_cluster: int, optional, default is 10
-            How many iterations of hill climbing to perform. More iterations
+        n_iterations_propagate_labels: int, optional, default is 30
+            Maximum number of iterations of label propagation to perform. More iterations
             will cluster more noise points.
 
-        n_neighbors_extend_cluster: int, optional, default is 5
+        n_neighbors_propagate_labels: int, optional, default is 5
             How many neighbors to use in the hill climbing procedure.
 
         returns:
@@ -157,7 +180,9 @@ class Persistable:
 
         """
         k = n_neighbors / self._mpspace._size
-        s = self._connection_radius * 2
+        default_percentile = 0.95
+        s = self._mpspace.connection_radius(default_percentile) * 2
+
         hc = self._mpspace.lambda_linkage([0, k], [s, 0])
         pd = hc.persistence_diagram()
         if pd.shape[0] == 0:
@@ -178,9 +203,9 @@ class Persistable:
[…]

[…]
