### 2023-01-09 741c363
move functionality in persistable class to mpspace
