+++ persistable/persistable.py
@@ -284,106 +284,6 @@ class Persistable:
             )
         return cl
 
-    def _find_end(self, fast=False):
-        if fast:
-            default_percentile = 0.95
-            return self._mpspace.connection_radius(default_percentile) * 4, self._maxk
-        else:
-            return self._mpspace.find_end()
-
-    def _compute_vineyard(self, start_end1, start_end2, n_parameters, n_jobs=4):
-        start1, end1 = start_end1
-        start2, end2 = start_end2
-        if (
-            start1[0] > end1[0]
-            or start1[1] < end1[1]
-            or start2[0] > end2[0]
-            or start2[1] < end2[1]
-        ):
-            raise ValueError(
-                "Parameters chosen for vineyard will result in non-monotonic lines!"
-            )
-        starts = list(
-            zip(
-                np.linspace(start1[0], start2[0], n_parameters),
-                np.linspace(start1[1], start2[1], n_parameters),
-            )
-        )
-        ends = list(
-            zip(
-                np.linspace(end1[0], end2[0], n_parameters),
-                np.linspace(end1[1], end2[1], n_parameters),
-            )
-        )
-        startends = list(zip(starts, ends))
-        pds = self._mpspace.lambda_linkage_vineyard(startends, n_jobs=n_jobs)
-        return Vineyard(startends, pds)
-
-    def _compute_hilbert_function(
-        self,
-        min_s,
-        max_s,
-        max_k,
-        min_k,
-        granularity,
-        n_jobs=4,
-    ):
-        if min_k >= max_k:
-            raise ValueError("min_k must be smaller than max_k.")
-        if min_s >= max_s:
-            raise ValueError("min_s must be smaller than max_s.")
-        if max_k > self._maxk:
-            max_k = min(max_k, self._maxk)
-            warnings.warn(
-                "Not enough neighbors to compute chosen max density threshold, using "
-                + str(max_k)
-                + " instead. If needed, re-initialize the Persistable instance with a larger n_neighbors."
-            )
-        if min_k >= max_k:
-            min_k = max_k / 2
-            warnings.warn(
-                "max density threshold too large, using " + str(min_k) + " instead."
-            )
-
-        ss = np.linspace(min_s, max_s, granularity)
-        ks = np.linspace(min_k, max_k, granularity)[::-1]
-        hf = self._mpspace.hilbert_function(ss, ks, n_jobs=n_jobs)
-        return ss, ks, hf, signed_betti(hf)
-
-    def _compute_rank_invariant(
-        self,
-        min_s,
-        max_s,
-        max_k,
-        min_k,
-        granularity,
-        reduced=False,
-        n_jobs=4,
-    ):
-        if min_k >= max_k:
-            raise ValueError("min_k must be smaller than max_k.")
-        if min_s >= max_s:
-            raise ValueError("min_s must be smaller than max_s.")
-        if max_k > self._maxk:
-            max_k = min(max_k, self._maxk)
-            warnings.warn(
-                "Not enough neighbors to compute chosen max density threshold, using "
-                + str(max_k)
-                + " instead. If needed, re-initialize the Persistable instance with a larger n_neighbors."
-            )
-        if min_k >= max_k:
-            min_k = max_k / 2
-            warnings.warn(
-                "max density threshold too large, using " + str(min_k) + " instead."
-            )
-
-        ss = np.linspace(min_s, max_s, granularity)
-        ks = np.linspace(min_k, max_k, granularity)[::-1]
-        ri = self._mpspace.rank_invariant(ss, ks, n_jobs=n_jobs, reduced=reduced)
[…]
