### 2023-01-09 0d70661
reflect refactoring in persistable_interactive
