### 2023-01-09 4a258e0
refactor construction of metric probability space
