### 2023-01-09 792c0e1
add RipsBifiltration abstraction:
