+++ persistable/persistable.py
@@ -138,6 +138,8 @@ class Persistable:
             X, metric, measure, self._n_neighbors, leaf_size, debug=debug, threading=threading, **kwargs
         )
 
+        self._bifiltration = _RipsBifiltration(self._mpspace)
+
     def quick_cluster(
         self,
         n_neighbors: int = 30,
@@ -181,9 +183,9 @@ class Persistable:
         """
         k = n_neighbors / self._mpspace._size
         default_percentile = 0.95
-        s = self._mpspace.connection_radius(default_percentile) * 2
+        s = self._bifiltration.connection_radius(default_percentile) * 2
 
-        hc = self._mpspace.lambda_linkage([0, k], [s, 0])
+        hc = self._bifiltration.lambda_linkage([0, k], [s, 0])
         pd = hc.persistence_diagram()
         if pd.shape[0] == 0:
             return np.full(self._mpspace._size, -1)
@@ -260,7 +262,7 @@ class Persistable:
             raise ValueError("start and end must both be points on the plane.")
         if n_clusters < 1:
             raise ValueError("n_clusters must be greater than 0.")
-        hc = self._mpspace.lambda_linkage(start, end)
+        hc = self._bifiltration.lambda_linkage(start, end)
         bd = hc.persistence_diagram()
         pers = np.abs(bd[:, 0] - bd[:, 1])
         # TODO: use sort from largest to smallest and make the logic below simpler
@@ -279,13 +281,13 @@ class Persistable:
             threshold = (spers[-n_clusters] + spers[-(n_clusters + 1)]) / 2
         cl = hc.persistence_based_flattening(threshold)
         if propagate_labels:
-            cl = self._mpspace._propagate_labels(
+            cl = self._mpspace.propagate_labels(
                 cl, n_iterations_propagate_labels, n_neighbors_propagate_labels
             )
         return cl
 
     def _find_end(self):
-        return self._mpspace.find_end()
+        return self._bifiltration.find_end()
 
     def _hilbert_function(
         self,
@@ -296,111 +298,31 @@ class Persistable:
         granularity,
         n_jobs=1,
     ):
-        return self._mpspace.hilbert_function_on_grid(
+        return self._bifiltration.hilbert_function_on_grid(
             min_s, max_s, max_k, min_k, granularity, n_jobs=n_jobs
         )
 
     def _rank_invariant(
         self, min_s, max_s, max_k, min_k, granularity, reduced=False, n_jobs=1
     ):
-        return self._mpspace.rank_invariant_on_grid(
+        return self._bifiltration.rank_invariant_on_grid(
             min_s, max_s, max_k, min_k, granularity, reduced=reduced, n_jobs=n_jobs
         )
 
     def _linear_vineyard(
         self, start_end1, start_end2, n_parameters, reduced=False, n_jobs=1
     ):
-        return self._mpspace.linear_vineyard(
+        return self._bifiltration.linear_vineyard(
             start_end1, start_end2, n_parameters, reduced=False, n_jobs=1
         )
 
 
-class _MetricProbabilitySpace:
-    """Implements a finite metric probability space that can compute its \
-       kernel density estimates and lambda linkage hierarchical clusterings """
-
-    _MAX_DIM_USE_BORUVKA = 60
+class _RipsBifiltration:
 
     def __init__(
-        self, X, metric, measure, n_neighbors, leaf_size=40, debug=False, threading=False, **kwargs
+        self, mpspace
     ):
-        # meta variables
-        self._threading = threading
-        self._debug = debug
-
-        # save extra arguments for metric
-        self._kwargs = kwargs
-
-        # save measure
-        self._measure = measure
-
-        # save point cloud
-        self._size = X.shape[0]
[…]
