+++ persistable/test/test_persistable.py
@@ -11,7 +11,7 @@ from sklearn import datasets
 import numpy as np
 
 
-class TestMetricProbabilitySpace(unittest.TestCase):
+class TestRipsBifiltration(unittest.TestCase):
     def setUp(self):
         np.random.seed(0)
         self._n = 25
@@ -30,44 +30,44 @@ class TestMetricProbabilitySpace(unittest.TestCase):
         X = np.random.random_sample((n, 2))
 
         p = Persistable(X, n_neighbors="all")
-        mps = p._mpspace
+        bf = p._bifiltration
         s0 = np.infty
         k0 = 1
-        mps._core_distance(np.arange(n), s0, k0)
+        bf._core_distance(np.arange(n), s0, k0)
 
     def test_core_distances(self):
         """ Check that the _core_distance method returns the correct answer """
         n = 4
         X = np.array([[0, 0], [1, 0], [1, 1], [3, 0]])
         p = Persistable(X)
-        mps = p._mpspace
+        bf = p._bifiltration
 
         s0 = 2
         k0 = 0.5
         res = np.array([1, 1, 1, 1])
-        np.testing.assert_almost_equal(mps._core_distance(np.arange(n), s0, k0), res)
+        np.testing.assert_almost_equal(bf._core_distance(np.arange(n), s0, k0), res)
 
         s0 = 1
         k0 = 0.5
         res = np.array([1 / 2, 1 / 2, 1 / 2, 1 / 2])
-        np.testing.assert_almost_equal(mps._core_distance(np.arange(n), s0, k0), res)
+        np.testing.assert_almost_equal(bf._core_distance(np.arange(n), s0, k0), res)
 
         s0 = np.infty
         k0 = 1
         res = np.array([3, 2, np.sqrt(5), 3])
-        np.testing.assert_almost_equal(mps._core_distance(np.arange(n), s0, k0), res)
+        np.testing.assert_almost_equal(bf._core_distance(np.arange(n), s0, k0), res)
 
         s0 = np.infty
         k0 = 0.5
         res = np.array([1, 1, 1, 2])
-        np.testing.assert_almost_equal(mps._core_distance(np.arange(n), s0, k0), res)
+        np.testing.assert_almost_equal(bf._core_distance(np.arange(n), s0, k0), res)
 
         p = Persistable(X, measure=np.array([0.5, 0.5, 0.5, 0.5]))
-        mps = p._mpspace
+        bf = p._bifiltration
         s0 = np.infty
         k0 = 0.6
         res = np.array([1, 1, 1, 2])
-        np.testing.assert_almost_equal(mps._core_distance(np.arange(n), s0, k0), res)
+        np.testing.assert_almost_equal(bf._core_distance(np.arange(n), s0, k0), res)
 
     #def test_core_distances_nonuniform_measure(self):
     #    """ Check that the _core_distance method returns the correct answer \
@@ -88,18 +88,18 @@ class TestMetricProbabilitySpace(unittest.TestCase):
         for w in self._different_weights:
             for p in self._ps:
                 p1 = Persistable(self._X, measure=w, p=p)
-                mps1 = p1._mpspace
+                bf1 = p1._bifiltration
                 p2 = Persistable(
                     distance_matrix(self._X, self._X, p=p),
                     metric="precomputed",
                     measure=w,
                 )
-                mps2 = p2._mpspace
+                bf2 = p2._bifiltration
                 for s0 in self._s0s:
                     for k0 in self._k0s:
                         np.testing.assert_almost_equal(
-                            mps1._core_distance(np.arange(self._n), s0, k0),
[…]
