### 2023-01-10 8932be7
simplify shifts in displaying of hilbert function and rank decomposition
