### 2023-01-11 8380b0d
refactor and creat MetricSpace class
