+++ persistable/persistable.py
@@ -181,14 +181,14 @@ class Persistable:
             i.e., points deemed not to belong to any cluster by the algorithm.
 
         """
-        k = n_neighbors / self._mpspace._size
+        k = n_neighbors / self._mpspace.size()
         default_percentile = 0.95
         s = self._bifiltration.connection_radius(default_percentile) * 2
 
         hc = self._bifiltration.lambda_linkage([0, k], [s, 0])
         pd = hc.persistence_diagram()
         if pd.shape[0] == 0:
-            return np.full(self._mpspace._size, -1)
+            return np.full(self._mpspace.size(), -1)
 
         def _prominences(bd):
             return np.sort(np.abs(bd[:, 0] - bd[:, 1]))[::-1]
@@ -281,7 +281,7 @@ class Persistable:
             threshold = (spers[-n_clusters] + spers[-(n_clusters + 1)]) / 2
         cl = hc.persistence_based_flattening(threshold)
         if propagate_labels:
-            cl = self._mpspace.propagate_labels(
+            cl = self._mpspace._mspace.propagate_labels(
                 cl, n_iterations_propagate_labels, n_neighbors_propagate_labels
             )
         return cl
@@ -298,14 +298,14 @@ class Persistable:
         granularity,
         n_jobs=1,
     ):
-        return self._bifiltration.hilbert_function_on_grid(
+        return self._bifiltration.hilbert_function_on_regular_grid(
             min_s, max_s, max_k, min_k, granularity, n_jobs=n_jobs
         )
 
     def _rank_invariant(
         self, min_s, max_s, max_k, min_k, granularity, reduced=False, n_jobs=1
     ):
-        return self._bifiltration.rank_invariant_on_grid(
+        return self._bifiltration.rank_invariant_on_regular_grid(
             min_s, max_s, max_k, min_k, granularity, reduced=reduced, n_jobs=n_jobs
         )
 
@@ -313,7 +313,7 @@ class Persistable:
         self, start_end1, start_end2, n_parameters, reduced=False, n_jobs=1
     ):
         return self._bifiltration.linear_vineyard(
-            start_end1, start_end2, n_parameters, reduced=False, n_jobs=1
+            start_end1, start_end2, n_parameters, reduced=reduced, n_jobs=n_jobs
         )
 
 
@@ -335,7 +335,7 @@ class _RipsBifiltration:
                 i_indices_and_finished_at_last_index.append(
                     lazy_intersection(
                         self._mpspace._kernel_estimate[p],
-                        self._mpspace._nn_distance[p],
+                        self._mpspace._mspace._nn_distance[p],
                         s_intercept,
                         k_intercept,
                     )
@@ -363,9 +363,9 @@ class _RipsBifiltration:
                     "Don't have enough neighbors to properly compute core scale, or point takes too long to appear."
                 )
             op = lambda p, i: np.where(
-                k_to_s(self._mpspace._kernel_estimate[p, i - 1]) <= self._mpspace._nn_distance[p, i],
+                k_to_s(self._mpspace._kernel_estimate[p, i - 1]) <= self._mpspace._mspace._nn_distance[p, i],
                 k_to_s(self._mpspace._kernel_estimate[p, i - 1]),
-                self._mpspace._nn_distance[p, i],
+                self._mpspace._mspace._nn_distance[p, i],
             )
             return np.where(i_indices == 0, 0, op(point_index, i_indices))
         else:
@@ -374,7 +374,7 @@ class _RipsBifiltration:
                 idx = np.searchsorted(
                     self._mpspace._kernel_estimate[p], k_intercept, side="left"
                 )
-                if idx == self._mpspace._nn_distance[p].shape[0]:
+                if idx == self._mpspace._mspace._nn_distance[p].shape[0]:
[…]
