### 2023-01-11 3e338b5
more modularization of metric space and metric probability space
