+++ persistable/persistable.py
@@ -135,15 +135,19 @@ class Persistable:
 
         # construct the filtration
         self._mpspace = _MetricProbabilitySpace(
-            X, metric, measure, self._n_neighbors, leaf_size, debug=debug, threading=threading, **kwargs
+            X, metric, measure, self._n_neighbors, leaf_size, **kwargs
         )
 
-        self._bifiltration = _RipsBifiltration(self._mpspace)
+        self._bifiltration = _RipsBifiltration(
+            self._mpspace,
+            debug=debug,
+            threading=threading,
+        )
 
     def quick_cluster(
         self,
         n_neighbors: int = 30,
-        n_clusters_range=[3, 15],
+        n_clusters_range=np.array([3, 15]),
         propagate_labels=False,
         n_iterations_propagate_labels=30,
         n_neighbors_propagate_labels=5,
@@ -281,7 +285,7 @@ class Persistable:
             threshold = (spers[-n_clusters] + spers[-(n_clusters + 1)]) / 2
         cl = hc.persistence_based_flattening(threshold)
         if propagate_labels:
-            cl = self._mpspace._mspace.propagate_labels(
+            cl = self._mpspace.propagate_labels(
                 cl, n_iterations_propagate_labels, n_neighbors_propagate_labels
             )
         return cl
@@ -296,10 +300,11 @@ class Persistable:
         max_k,
         min_k,
         granularity,
+        reduced=False,
         n_jobs=1,
     ):
         return self._bifiltration.hilbert_function_on_regular_grid(
-            min_s, max_s, max_k, min_k, granularity, n_jobs=n_jobs
+            min_s, max_s, max_k, min_k, granularity, reduced=reduced, n_jobs=n_jobs
         )
 
     def _rank_invariant(
@@ -318,13 +323,16 @@ class Persistable:
 
 
 class _RipsBifiltration:
-
-    def __init__(
-        self, mpspace
-    ):
+    def __init__(self, mpspace, debug=False, threading=False):
+        self._debug = debug
+        self._threading = threading
         self._mpspace = mpspace
 
     def _core_distance(self, point_index, s_intercept, k_intercept, max_k=None):
+
+        kernel_estimate = self._mpspace.kernel_estimate()
+        nn_distance = self._mpspace.nn_distance()
+
         max_k = k_intercept if max_k is None else max_k
         if s_intercept != np.inf:
             i_indices_and_finished_at_last_index = []
@@ -334,14 +342,14 @@ class _RipsBifiltration:
             for p in point_index:
                 i_indices_and_finished_at_last_index.append(
                     lazy_intersection(
-                        self._mpspace._kernel_estimate[p],
-                        self._mpspace._mspace._nn_distance[p],
+                        kernel_estimate[p],
+                        nn_distance[p],
                         s_intercept,
                         k_intercept,
                     )
                 )
                 max_k_larger_last_kernel_estimate.append(
-                    (self._mpspace._kernel_estimate[p, -1] < max_k)
+                    (kernel_estimate[p, -1] < max_k)
                 )
             max_k_larger_last_kernel_estimate = np.array(
                 max_k_larger_last_kernel_estimate
@@ -362,19 +370,20 @@ class _RipsBifiltration:
                 warnings.warn(
                     "Don't have enough neighbors to properly compute core scale, or point takes too long to appear."
                 )
-            op = lambda p, i: np.where(
-                k_to_s(self._mpspace._kernel_estimate[p, i - 1]) <= self._mpspace._mspace._nn_distance[p, i],
-                k_to_s(self._mpspace._kernel_estimate[p, i - 1]),
-                self._mpspace._mspace._nn_distance[p, i],
-            )
+
[…]
