### 2023-01-11 e504d98
add todo for persistent metric space
