+++ persistable/subsampling.pyx
@@ -12,10 +12,49 @@ from libc.float cimport DBL_MAX
 from borrowed.dist_metrics cimport DistanceMetric
 
 
-#cpdef np.ndarray[np.int64_t, ndim=1] close_subsample_distance_matrix(subsample_size, seed=0):
+cpdef close_subsample_distance_matrix(
+    np.int64_t subsample_size,
+    np.ndarray[np.double_t, ndim=2, mode='c'] dist_mat,
+    random_start=0
+):
+    cdef np.int64_t dim = dist_mat.shape[0]
+
+    cdef np.double_t[:, ::1] dist_mat_view = dist_mat
+
+    cdef np.double_t val1
+    cdef np.double_t val2
+
+    cdef np.int64_t idx
+
+    idx_perm = np.zeros(subsample_size, dtype=np.int64)
+    cdef np.int64_t[:] idx_perm_view = idx_perm
+    radii = np.zeros(subsample_size)
+    cdef np.double_t[:] radii_view = radii
+
+    cdef np.double_t[:] ds = np.full(dim, np.inf)
+
+    idx = random_start
+    idx_perm_view[0] = idx
+    for j in range(0, dim):
+        val1 = ds[j]
+        val2 = dist_mat[idx,j]
+        ds[j] = min(val1, val2)
+
+    for i in range(1, subsample_size):
+        idx = np.argmax(ds)
+        idx_perm_view[i] = idx
+        radii_view[i - 1] = ds[idx]
+        for j in range(0, dim):
+            val1 = ds[j]
+            val2 = dist_mat[idx,j]
+            ds[j] = min(val1, val2)
+
+    radii_view[-1] = np.max(ds)
+
+    return idx_perm, radii
 
 
-cpdef np.ndarray[np.int64_t, ndim=1] close_subsample_fast_metric(
+cpdef close_subsample_fast_metric(
     np.int64_t subsample_size,
     np.ndarray[np.double_t, ndim=2, mode='c'] raw_data,
     DistanceMetric dist_metric,
@@ -38,7 +77,8 @@ cpdef np.ndarray[np.int64_t, ndim=1] close_subsample_fast_metric(
 
     idx_perm = np.zeros(subsample_size, dtype=np.int64)
     cdef np.int64_t[:] idx_perm_view = idx_perm
-    cdef np.double_t[:] radii = np.zeros(subsample_size)
+    radii = np.zeros(subsample_size)
+    cdef np.double_t[:] radii_view = radii
 
     cdef np.double_t[:] ds = np.full(dim, np.inf)
 
@@ -54,7 +94,7 @@ cpdef np.ndarray[np.int64_t, ndim=1] close_subsample_fast_metric(
     for i in range(1, subsample_size):
         idx = np.argmax(ds)
         idx_perm_view[i] = idx
-        radii[i - 1] = ds[idx]
+        radii_view[i - 1] = ds[idx]
         for j in range(0, dim):
             val1 = ds[j]
             val2 = dist_metric.dist(&raw_data_ptr[num_features * idx],
@@ -62,71 +102,7 @@ cpdef np.ndarray[np.int64_t, ndim=1] close_subsample_fast_metric(
                                      num_features)
             ds[j] = min(val1, val2)
 
-    radii[-1] = np.max(ds)
+    radii_view[-1] = np.max(ds)
 
-    return idx_perm
-    #return idx_perm, radii
+    return idx_perm, radii
 
-#        #for i in range(1, dim):
-#        #    in_tree[current_node] = 1
-#    
-#        #    current_node_core_distance = core_distance_ptr[current_node]
-#    
-#        #    new_distance = DBL_MAX
-#        #    source_node = 0
-#        #    new_node = 0
-#    
[…]
