### 2023-01-11 dad0a0b
add subsampling option to user-facing persistable class
