+++ persistable/persistable.py
@@ -74,6 +74,9 @@ class Persistable:
         the number of points in the dataset, if set to ``"auto"`` it will find
         a reasonable default.
 
+    subsample: None or int, optional, default is None
+        Number of datapoints to do measure-consistent subsampling.
+
     debug: bool, optional, default is False
         Whether to print debug messages.
 
@@ -92,51 +95,81 @@ class Persistable:
         self,
         X,
         metric="minkowski",
+        measure=None,
         n_neighbors="auto",
+        subsample=None,
         debug=False,
         threading=False,
         **kwargs
     ):
-        self._debug = debug
-        self._threading = threading
-        # keep dataset
-        self._data = X
-        # if metric is minkowski but no p was passed, assume p = 2
-        if metric == "minkowski" and "p" not in kwargs:
-            kwargs["p"] = 2
-        # if no measure was passed, assume normalized counting measure
-        if "measure" not in kwargs:
-            measure = np.full(X.shape[0], 1.0 / X.shape[0])
-        elif "measure" in kwargs and kwargs["measure"] is None:
-            measure = np.full(X.shape[0], 1.0 / X.shape[0])
-            del kwargs["measure"]
-        else:
-            measure = kwargs["measure"]
-            del kwargs["measure"]
+        self._subsample = None 
+
         if "leaf_size" in kwargs:
             leaf_size = kwargs["leaf_size"]
         else:
             leaf_size = 40
+
+        # if no measure was passed, assume normalized counting measure
+        if measure is None:
+            measure = np.full(X.shape[0], 1.0 / X.shape[0])
+        else:
+            measure = np.array(measure)
+            assert np.all(measure >= 0)
+            # rescale so that total measure is 1
+            measure = measure / np.sum(measure)
+
+        if subsample is not None:
+            if type(subsample) != int:
+                raise ValueError("subsample must be either None or an integer.")
+            ms = _MetricSpace(X, metric, **kwargs)
+            if subsample >= X.shape[0]:
+                subsample = X.shape[0] / 10
+                warnings.warn(
+                    "subsample is greater than or equal to number of datapoints, using "
+                    + str(subsample)
+                    + " instead."
+                )
+            self._subsample = subsample
+            subsample_indices, _, subsample_representatives = ms.close_subsample(
+                subsample
+            )
+            X = X.copy()
+            if metric == "precomputed":
+                X = X[subsample_indices, :][:, subsample_indices]
+            else:
+                X = X[subsample_indices, :]
+            self._subsample_representatives = subsample_representatives
+
+            # compute measure for subsample
+            new_measure = np.zeros(subsample)
+            for i, _ in enumerate(self._subsample_representatives):
+                new_measure[self._subsample_representatives[i]] += measure[i]
+            measure = new_measure
+
+        # if metric is minkowski but no p was passed, assume p = 2
+        if metric == "minkowski" and "p" not in kwargs:
+            kwargs["p"] = 2
+
         # if no n_neighbors for fitting mpspace was passed, compute a reasonable one
         if n_neighbors == "auto":
             if X.shape[0] < 100:
-                self._n_neighbors = X.shape[0]
+                n_neighbors = X.shape[0]
             else:
-                self._n_neighbors = min(int(np.log10(X.shape[0])) * 100, X.shape[0])
+                n_neighbors = min(int(np.log10(X.shape[0])) * 100, X.shape[0])
         elif n_neighbors == "all":
-            self._n_neighbors = X.shape[0]
+            n_neighbors = X.shape[0]
[…]
