### 2023-01-11 e1648f4
add measure input description to persistable init
