### 2023-01-11 0757c91
delete obsolete class constant
