### 2023-01-13 #5022 @AP6YC
Hi @rkurchin, sorry for the delay. I will get my checklist started below and begin my review.
