### 2023-01-23 #5022 @lmcinnes
Sorry, I was having installation issues (on my end from my environment manager; not the package), and set it aside for a little while.

I have a couple of minor issues that I'll file imminently, but mostly it looks good to me.
