### 2023-01-23 LuisScoccola/persistable#19 @lmcinnes
Dash warnings in ``start_UI(inline=True)``
When running in a notebook on a clean install the ``start_UI(inline=True)`` works, but results in a large number of warnings being printed to the notebook from Dash. You may want to find a way to catch or suppress these warnings as they can be distracting.

openjournals/joss-reviews#5022
