### 2023-01-23 LuisScoccola/persistable#20 @lmcinnes
Installation requirements
The installation went very smoothly, relying on pip to pull down all the requirements. It would be helpful, however, to list the top level installation requirements in the documentation up front so that users can have a clear idea of what they will be pulling in before they do the install (I know this is covered by ``requirements.txt`` but putting it in the README and readthedocs would be beneficial).

#openjournals/joss-reviews#5022
