### 2023-01-23 LuisScoccola/persistable#21 @lmcinnes
Community guidelines
The Contributing section lists what to do for people wishing to contribute, but it would be beneficial if you could also cover those seeking help with issues (e.g. how to file an issue, and what the expectations are for issue formatting if any), and simply with questions relating to the library (e.g. a discussions section). As a further step it could be helpful to include a code of conduct for the repository, just to make expectations clear.

openjournals/joss-reviews#5022
