### 2023-01-23 9cd0ab8
supress warning messages when using inline mode not in debug mode
