### 2023-01-23 9b09d7f
fix: issue 19 (#22)
Fix #19 where running persistable in inline mode and not in debug mode would generate a large amount of warnings and messages.
