### 2023-01-23 6f47183
describe persistable dependencies in readme and documentation
