### 2023-01-23 2bf8407
fix rendering of bash in documentation
