### 2023-01-23 b43c466
second try to fix rendering of bash in documentation
