+++ docs/index.rst
@@ -16,7 +16,7 @@ Installing
 Make sure you are using Python 3.
 Installation through PyPI is coming soon.
 
-.. code:: bash
+.. code-block:: none
     pip install git+https://github.com/LuisScoccola/persistable.git
 
 Persistable depends on the following python packages:
