### 2023-01-23 8157e3e
third try to fix rendering of bash in documentation
