### 2023-01-23 4a5a6e2
fourth try to fix rendering of bash in documentation
