### 2023-01-24 5b75933
docs: issue 20 (#23)
Address #20 by including Persistable's dependencies in the documentation and in the readme.
