### 2023-01-24 016a0ac
include more info in readme about issues
