### 2023-01-24 c946d2a
docs: issue 21 (#24)
Address #21 by adding a code of conduct and more information about how to file an issue to ask questions.
