### 2023-01-25 9e51b3c
add toggle for one-parameter slice selection
