+++ persistable/persistable_interactive.py
@@ -45,7 +45,7 @@ Y_START_SECOND_LINE = "y-start-second-line-"
 X_END_SECOND_LINE = "x-end-second-line-"
 Y_END_SECOND_LINE = "y-end-second-line-"
 CCF_PLOT = "ccf-plot-"
-DISPLAY_LINES_SELECTION = "display-lines-selection-"
+INTERACTIVE_INPUTS_SELECTION = "interactive-inputs-selection-"
 ENDPOINT_SELECTION = "endpoint-selection-"
 STORED_CCF = "stored-ccf-"
 STORED_X_TICKS_CCF = "stored-x-ticks-ccf-"
@@ -82,8 +82,10 @@ INPUT_REDUCED_HOMOLOGY_RI = "input-reduced-homology-ri-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
 CCF_EXTRAS = "ccf-extras-"
+PV_PANEL = "pv-panel-"
 PV_DETAILS = "pv-details-"
 PV_PLOT_CONTROLS_DIV = "pv-plot-controls-div-"
+PD_PANEL = "pd-panel-"
 LOG = "log-"
 LOG_DIV = "log-div-"
 STORED_PV = "stored-pv-"
@@ -92,6 +94,7 @@ INPUT_PROM_VIN_SCALE = "input-prom-vin-scale-"
 COMPUTE_PV_BUTTON = "compute-pv-button-"
 STOP_COMPUTE_PV_BUTTON = "stop-compute-pv-button-"
 PV_PLOT = "pv-plot-"
+PD_PLOT = "pd-plot-"
 STORED_PV_DRAWING = "stored-pv-drawing-"
 INPUT_GRANULARITY_PV = "input-granularity-pv-"
 INPUT_NUM_JOBS_PV = "input-num-jobs-pv-"
@@ -463,12 +466,16 @@ class PersistableInteractive:
                                         children=[
                                             html.Span(
                                                 className="name",
-                                                children="Vineyard inputs selection",
+                                                children="Interactive inputs selection",
                                             ),
                                             dcc.RadioItems(
-                                                ["On", "Off"],
+                                                [
+                                                    "Off",
+                                                    "One-parameter slice",
+                                                    "Line family",
+                                                ],
                                                 "Off",
-                                                id=DISPLAY_LINES_SELECTION,
+                                                id=INTERACTIVE_INPUTS_SELECTION,
                                                 className=VALUE,
                                             ),
                                         ],
@@ -945,70 +952,126 @@ class PersistableInteractive:
                 #
                 html.Div(id=EXPORTED_PARAMETER, hidden=True),
                 html.Div(
-                    className="grid",
+                    className="horizontal-grid",
                     children=[
                         html.Div(
+                            className="vertical-grid",
                             children=[
-                                html.H2("Component Counting Function"),
+                                html.Div(
+                                    children=[
+                                        html.H2("Component Counting Function"),
+                                        html.Div(
+                                            className="parameters",
+                                            children=ccf_inputs + ccf_buttons,
+                                        ),
+                                    ]
+                                ),
+                                dcc.Graph(
+                                    id=CCF_PLOT,
+                                    className="graph",
+                                    figure=empty_figure(),
+                                    config={
+                                        "responsive": True,
+                                        "displayModeBar": False,
+                                        "modeBarButtonsToRemove": [
+                                            "toImage",
+                                            "pan",
+                                            "zoomIn",
+                                            "zoomOut",
+                                            "resetScale",
+                                            "lasso",
+                                            "select",
+                                        ],
+                                        "displaylogo": False,
+                                    },
+                                ),
                                 html.Div(
                                     className="parameters",
[…]
