### 2023-01-25 0c35815
add timeout to ui tests so that windows tests don't fail sometimes
