### 2023-01-25 39ce1b6
add inputs for persistence diagram choice
