### 2023-01-25 4a0dbca
add interactivity to one-parameter slice selection
