+++ persistable/persistable_interactive.py
@@ -50,7 +50,8 @@ X_END_SECOND_LINE = "x-end-second-line-"
 Y_END_SECOND_LINE = "y-end-second-line-"
 CCF_PLOT = "ccf-plot-"
 INTERACTIVE_INPUTS_SELECTION = "interactive-inputs-selection-"
-ENDPOINT_SELECTION = "endpoint-selection-"
+PV_ENDPOINT_SELECTION = "pv-endpoint-selection-"
+PD_ENDPOINT_SELECTION = "pd-endpoint-selection-"
 STORED_CCF = "stored-ccf-"
 STORED_X_TICKS_CCF = "stored-x-ticks-ccf-"
 STORED_Y_TICKS_CCF = "stored-y-ticks-ccf-"
@@ -64,9 +65,11 @@ MIN_DIST_SCALE = "min-dist-scale-"
 MAX_DIST_SCALE = "max-dist-scale-"
 MIN_DENSITY_THRESHOLD = "min-density-threshold-"
 MAX_DENSITY_THRESHOLD = "max-density-threshold-"
-ENDPOINT_SELECTION_DIV = "endpoint-selection-div-"
+PV_ENDPOINT_SELECTION_DIV = "pv-endpoint-selection-div-"
+PD_ENDPOINT_SELECTION_DIV = "pd-endpoint-selection-div-"
 PARAMETER_SELECTION_DIV = "parameter-selection-div-"
 DISPLAY_PARAMETER_SELECTION = "display-parameter-selection-"
+DISPLAY_SELECTED_LINE = "display-selected-line-"
 COMPUTE_CCF_BUTTON = "compute-ccf-button-"
 STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
 COMPUTE_RI_BUTTON = "compute-ri-button-"
@@ -108,7 +111,7 @@ INPUT_NUM_JOBS_PV = "input-num-jobs-pv-"
 INPUT_LINE = "input-line-"
 INPUT_GAP = "gap-"
 EXPORT_PARAMETERS_BUTTON = "export-parameters-"
-FIXED_PARAMETERS = "fixed-parameters-"
+PV_FIXED_PARAMETERS = "fixed-parameters-"
 STORED_PD = "stored-pd-"
 
 STORED_CCF_COMPUTATION_WARNINGS = "stored-ccf-computation-warnings-"
@@ -303,6 +306,14 @@ class PersistableInteractive:
             (defr - 1) / defr
         )
         default_y_end_second_line = (default_min_k + default_max_k) * (1 / 2)
+        default_x_start_line = (
+            default_x_start_first_line + default_x_start_second_line
+        ) / 2
+        default_y_start_line = (
+            default_y_start_first_line + default_y_start_second_line
+        ) / 2
+        default_x_end_line = (default_x_end_first_line + default_x_end_second_line) / 2
+        default_y_end_line = (default_y_end_first_line + default_y_end_second_line) / 2
 
         self._app.title = "Persistable"
 
@@ -468,6 +479,24 @@ class PersistableInteractive:
                             html.Div(
                                 className="parameters",
                                 children=[
+                                    # html.Div(
+                                    #    className="parameter-single",
+                                    #    children=[
+                                    #        html.Span(
+                                    #            className="name",
+                                    #            children="Display selected line",
+                                    #        ),
+                                    #        dcc.RadioItems(
+                                    #            [
+                                    #                "On",
+                                    #                "Off",
+                                    #            ],
+                                    #            "On",
+                                    #            id=DISPLAY_SELECTED_LINE,
+                                    #            className=VALUE,
+                                    #        ),
+                                    #    ],
+                                    # ),
                                     html.Div(
                                         className="parameter-single",
                                         children=[
[…]
