### 2023-01-25 0dd8953
add computation and plot of barcode
