+++ persistable/persistable_interactive.py
@@ -112,11 +112,13 @@ INPUT_LINE = "input-line-"
 INPUT_GAP = "gap-"
 EXPORT_PARAMETERS_BUTTON = "export-parameters-"
 PV_FIXED_PARAMETERS = "fixed-parameters-"
-STORED_PD = "stored-pd-"
+STORED_PD_BY_PV = "stored-pd-by-pv-"
+STORED_PARAMETERS_AND_PD_BY_PD = "stored-pd-by-pd-"
 
 STORED_CCF_COMPUTATION_WARNINGS = "stored-ccf-computation-warnings-"
 STORED_PV_COMPUTATION_WARNINGS = "stored-pv-computation-warnings-"
 STORED_RI_COMPUTATION_WARNINGS = "stored-ri-computation-warnings-"
+STORED_PD_COMPUTATION_WARNINGS = "stored-pd-computation-warnings-"
 
 EXPORTED_PARAMETER = "exported-parameter-"
 
@@ -1089,11 +1091,13 @@ class PersistableInteractive:
                 # contains the basic prominence vineyard plot as a plotly figure
                 dcc.Store(id=STORED_PV_DRAWING),
                 #
-                dcc.Store(id=STORED_PD, data=json.dumps([])),
+                dcc.Store(id=STORED_PD_BY_PV, data=json.dumps([])),
+                dcc.Store(id=STORED_PARAMETERS_AND_PD_BY_PD, data=json.dumps([])),
                 #
                 dcc.Store(id=STORED_CCF_COMPUTATION_WARNINGS, data=json.dumps(" ")),
                 dcc.Store(id=STORED_RI_COMPUTATION_WARNINGS, data=json.dumps(" ")),
                 dcc.Store(id=STORED_PV_COMPUTATION_WARNINGS, data=json.dumps(" ")),
+                dcc.Store(id=STORED_PD_COMPUTATION_WARNINGS, data=json.dumps(" ")),
                 #
                 dcc.Store(id=PV_FIXED_PARAMETERS, data=json.dumps([])),
                 #
@@ -1354,6 +1358,7 @@ class PersistableInteractive:
                 [STORED_CCF_COMPUTATION_WARNINGS, DATA, IN],
                 [STORED_RI_COMPUTATION_WARNINGS, DATA, IN],
                 [STORED_PV_COMPUTATION_WARNINGS, DATA, IN],
+                [STORED_PD_COMPUTATION_WARNINGS, DATA, IN],
             ],
             [[LOG, CHILDREN], [LOG_DIV, "open"]],
             True,
@@ -1365,6 +1370,8 @@ class PersistableInteractive:
                 message = json.loads(d[STORED_PV_COMPUTATION_WARNINGS + DATA])
             elif ctx.triggered_id == STORED_RI_COMPUTATION_WARNINGS:
                 message = json.loads(d[STORED_RI_COMPUTATION_WARNINGS + DATA])
+            elif ctx.triggered_id == STORED_PD_COMPUTATION_WARNINGS:
+                message = json.loads(d[STORED_PD_COMPUTATION_WARNINGS + DATA])
             else:
                 raise Exception(
                     "print_log was triggered by unknown id: " + str(ctx.triggered_id)
@@ -1567,7 +1574,8 @@ class PersistableInteractive:
                 [DISPLAY_PARAMETER_SELECTION, VALUE, IN],
                 [INTERACTIVE_INPUTS_SELECTION, VALUE, IN],
                 [PV_FIXED_PARAMETERS, DATA, IN],
-                [STORED_PD, DATA, ST],
+                [STORED_PD_BY_PV, DATA, ST],
+                [STORED_PARAMETERS_AND_PD_BY_PD, DATA, IN],
[…]
