### 2023-01-25 9df020f
add basic persistence diagram
