### 2023-01-25 d27d238
clean pd plot
