+++ persistable/persistable_interactive.py
@@ -67,8 +67,10 @@ MIN_DENSITY_THRESHOLD = "min-density-threshold-"
 MAX_DENSITY_THRESHOLD = "max-density-threshold-"
 PV_ENDPOINT_SELECTION_DIV = "pv-endpoint-selection-div-"
 PD_ENDPOINT_SELECTION_DIV = "pd-endpoint-selection-div-"
-PARAMETER_SELECTION_DIV = "parameter-selection-div-"
-DISPLAY_PARAMETER_SELECTION = "display-parameter-selection-"
+PARAMETER_SELECTION_DIV_PV = "parameter-selection-div-pv-"
+PARAMETER_SELECTION_DIV_PD = "parameter-selection-div-pd-"
+DISPLAY_PARAMETER_SELECTION_PV = "display-parameter-selection-pv-"
+DISPLAY_PARAMETER_SELECTION_PD = "display-parameter-selection-pd-"
 DISPLAY_SELECTED_LINE = "display-selected-line-"
 COMPUTE_CCF_BUTTON = "compute-ccf-button-"
 STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
@@ -93,6 +95,7 @@ PV_PANEL = "pv-panel-"
 PV_DETAILS = "pv-details-"
 PD_DETAILS = "pd-details-"
 PV_PLOT_CONTROLS_DIV = "pv-plot-controls-div-"
+PD_PLOT_CONTROLS_DIV = "pd-plot-controls-div-"
 PD_PANEL = "pd-panel-"
 LOG = "log-"
 LOG_DIV = "log-div-"
@@ -110,7 +113,9 @@ INPUT_GRANULARITY_PV = "input-granularity-pv-"
 INPUT_NUM_JOBS_PV = "input-num-jobs-pv-"
 INPUT_LINE = "input-line-"
 PV_INPUT_GAP = "pv-input-gap-"
-EXPORT_PARAMETERS_BUTTON = "export-parameters-"
+PD_INPUT_GAP = "pd-input-gap-"
+EXPORT_PARAMETERS_BUTTON_PV = "export-parameters-button-pv-"
+EXPORT_PARAMETERS_BUTTON_PD = "export-parameters-button-pd-"
 PV_FIXED_PARAMETERS = "fixed-parameters-"
 STORED_PD_BY_PV = "stored-pd-by-pv-"
 STORED_PARAMETERS_AND_PD_BY_PD = "stored-pd-by-pd-"
@@ -481,24 +486,6 @@ class PersistableInteractive:
                             html.Div(
                                 className="parameters",
                                 children=[
-                                    # html.Div(
-                                    #    className="parameter-single",
-                                    #    children=[
-                                    #        html.Span(
-                                    #            className="name",
-                                    #            children="Display selected line",
-                                    #        ),
-                                    #        dcc.RadioItems(
-                                    #            [
-                                    #                "On",
-                                    #                "Off",
-                                    #            ],
-                                    #            "On",
-                                    #            id=DISPLAY_SELECTED_LINE,
-                                    #            className=VALUE,
-                                    #        ),
-                                    #    ],
-                                    # ),
                                     html.Div(
                                         className="parameter-single",
                                         children=[
@@ -933,14 +920,14 @@ class PersistableInteractive:
                             dcc.RadioItems(
                                 ["On", "Off"],
                                 "Off",
-                                id=DISPLAY_PARAMETER_SELECTION,
+                                id=DISPLAY_PARAMETER_SELECTION_PV,
                                 className=VALUE,
                             ),
                         ],
                     ),
                     html.Div(
                         className="parameter-double-button",
-                        id=PARAMETER_SELECTION_DIV,
[…]
