### 2023-01-25 61ac6d7
fix inconsistent name in test after updates
