### 2023-01-25 90fc217
fix inconsistent name in test after updates
