### 2023-01-25 40c7e39
feat: add parameter selection using a single line (#25)
* add toggle for one-parameter slice selection

* add timeout to ui tests so that windows tests don't fail sometimes

* add inputs for persistence diagram choice

* add interactivity to one-parameter slice selection

* add persistence diagram window
