+++ persistable/persistable_interactive.py
@@ -36,6 +36,10 @@ BaseLongCallbackManager.hash_function = monkeypatched_hash_function
 ###
 
 
+X_START_LINE = "x-start-line-"
+Y_START_LINE = "y-start-line-"
+X_END_LINE = "x-end-line-"
+Y_END_LINE = "y-end-line-"
 X_START_FIRST_LINE = "x-start-first-line-"
 Y_START_FIRST_LINE = "y-start-first-line-"
 X_END_FIRST_LINE = "x-end-first-line-"
@@ -45,8 +49,9 @@ Y_START_SECOND_LINE = "y-start-second-line-"
 X_END_SECOND_LINE = "x-end-second-line-"
 Y_END_SECOND_LINE = "y-end-second-line-"
 CCF_PLOT = "ccf-plot-"
-DISPLAY_LINES_SELECTION = "display-lines-selection-"
-ENDPOINT_SELECTION = "endpoint-selection-"
+INTERACTIVE_INPUTS_SELECTION = "interactive-inputs-selection-"
+PV_ENDPOINT_SELECTION = "pv-endpoint-selection-"
+PD_ENDPOINT_SELECTION = "pd-endpoint-selection-"
 STORED_CCF = "stored-ccf-"
 STORED_X_TICKS_CCF = "stored-x-ticks-ccf-"
 STORED_Y_TICKS_CCF = "stored-y-ticks-ccf-"
@@ -60,9 +65,13 @@ MIN_DIST_SCALE = "min-dist-scale-"
 MAX_DIST_SCALE = "max-dist-scale-"
 MIN_DENSITY_THRESHOLD = "min-density-threshold-"
 MAX_DENSITY_THRESHOLD = "max-density-threshold-"
-ENDPOINT_SELECTION_DIV = "endpoint-selection-div-"
-PARAMETER_SELECTION_DIV = "parameter-selection-div-"
-DISPLAY_PARAMETER_SELECTION = "display-parameter-selection-"
+PV_ENDPOINT_SELECTION_DIV = "pv-endpoint-selection-div-"
+PD_ENDPOINT_SELECTION_DIV = "pd-endpoint-selection-div-"
+PARAMETER_SELECTION_DIV_PV = "parameter-selection-div-pv-"
+PARAMETER_SELECTION_DIV_PD = "parameter-selection-div-pd-"
+DISPLAY_PARAMETER_SELECTION_PV = "display-parameter-selection-pv-"
+DISPLAY_PARAMETER_SELECTION_PD = "display-parameter-selection-pd-"
+DISPLAY_SELECTED_LINE = "display-selected-line-"
 COMPUTE_CCF_BUTTON = "compute-ccf-button-"
 STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
 COMPUTE_RI_BUTTON = "compute-ri-button-"
@@ -82,8 +91,12 @@ INPUT_REDUCED_HOMOLOGY_RI = "input-reduced-homology-ri-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
 CCF_EXTRAS = "ccf-extras-"
+PV_PANEL = "pv-panel-"
 PV_DETAILS = "pv-details-"
+PD_DETAILS = "pd-details-"
 PV_PLOT_CONTROLS_DIV = "pv-plot-controls-div-"
+PD_PLOT_CONTROLS_DIV = "pd-plot-controls-div-"
+PD_PANEL = "pd-panel-"
 LOG = "log-"
 LOG_DIV = "log-div-"
 STORED_PV = "stored-pv-"
[…]
