### 2023-01-26 3e38434
use different trees for nn_fit and for spanning tree computation
