+++ persistable/persistable.py
@@ -94,6 +94,10 @@ class Persistable:
         significantly slower because of the GIL, but the backend ``loky`` does
         not work well in some systems.
 
+    n_jobs: int, default is 1
+        Number of processes or threads to use to fit the data structures, for exaple
+        to compute the nearest neighbors of all points in the dataset.
+
     ``**kwargs``:
         Passed to ``KDTree`` or ``BallTree``.
 
@@ -108,14 +112,10 @@ class Persistable:
         n_neighbors="auto",
         debug=False,
         threading=False,
+        n_jobs=4,
         **kwargs
     ):
         self._subsample = None 
-
-        if "leaf_size" in kwargs:
-            leaf_size = kwargs["leaf_size"]
-        else:
-            leaf_size = 40
 
         # if no measure was passed, assume normalized counting measure
         if measure is None:
@@ -129,7 +129,7 @@ class Persistable:
         if subsample is not None:
             if type(subsample) != int:
                 raise ValueError("subsample must be either None or an integer.")
-            ms = _MetricSpace(X, metric, **kwargs)
+            ms = _MetricSpace(X, metric, threading=threading, debug=debug, n_jobs=n_jobs, **kwargs)
             if subsample >= X.shape[0]:
                 subsample = int(X.shape[0] / 4)
                 warnings.warn(
@@ -177,7 +177,7 @@ class Persistable:
 
         # construct the filtration
         self._mpspace = _MetricProbabilitySpace(
-            X, metric, measure, n_neighbors, leaf_size, **kwargs
+            X, metric, measure, n_neighbors, threading=threading, debug=debug, n_jobs=n_jobs, **kwargs
         )
 
         self._bifiltration = _DegreeRipsBifiltration(
@@ -192,7 +192,7 @@ class Persistable:
         n_clusters_range=np.array([3, 15]),
         propagate_labels=False,
         n_iterations_propagate_labels=30,
-        n_neighbors_propagate_labels=5,
+        n_neighbors_propagate_labels=15,
     ):
         """Clusters the dataset with which the Persistable instance was initialized.
 
@@ -263,7 +263,7 @@ class Persistable:
         end,
         propagate_labels=False,
         n_iterations_propagate_labels=30,
-        n_neighbors_propagate_labels=5,
+        n_neighbors_propagate_labels=15,
     ):
         """Clusters the dataset with which the Persistable instance was initialized.
 
@@ -339,8 +339,8 @@ class Persistable:
 
         return cl
 
-    def _find_end(self):
-        return self._bifiltration.find_end()
+    def _find_end(self, fast):
+        return self._bifiltration.find_end(fast=fast)
 
     def _hilbert_function(
         self,
@@ -792,9 +792,13 @@ class _MetricSpace:
 
     _MAX_DIM_USE_BORUVKA = 60
 
-    def __init__(self, X, metric, leaf_size=40, **kwargs):
+    def __init__(self, X, metric, leaf_size=40, threading=False, debug=False, n_jobs=1, **kwargs):
         # save extra arguments for metric
         self._kwargs = kwargs
+
+        self._threading = threading
+        self._n_jobs = n_jobs
+        self._debug = debug
 
         # default values before fitting
         self._fitted_nn = False
@@ -805,7 +809,8 @@ class _MetricSpace:
         self._size = None
         self._dimension = None
         self._points = None
-        self._tree = None
+        self._nn_tree = None
+        self._boruvka_tree = None
         self._dist_mat = None
         self._dist_metric = None
 
@@ -824,13 +829,20 @@ class _MetricSpace:
         self._metric = metric
         self._leaf_size = leaf_size
         if metric in KDTree.valid_metrics + BallTree.valid_metrics:
+            leaf_size_boruvka = 3 if self._leaf_size < 3 else self._leaf_size // 3
[…]
