### 2023-01-26 dc04320
don't handle dataset size in persistable_interactive
