### 2023-01-27 6ca2a20
better defaults for granularity for large data
