### 2023-01-27 22cd300
add fast subsampling for euclidean data
