### 2023-01-27 2e4c5c7
handle euclidean subsample a bit better
