### 2023-01-27 e1f44b4
fix problem when subsampling routine does not output the exact number of requested samples
