### 2023-01-28 ea7b67b
perf: use two different metric trees for nn and mst, and add faster subsampling for Euclidean data (#26)
* use different trees for nn_fit and for spanning tree computation

* better defaults for granularity for large data

* add fast subsampling for euclidean data
