+++ persistable/persistable.py
@@ -31,7 +31,8 @@ from joblib.parallel import cpu_count
 
 
 _TOL = 1e-08
-
+# starting when we consider a dataset large
+_MANY_POINTS = 16000
 
 def parallel_computation(function, inputs, n_jobs, debug=False, threading=False):
     if n_jobs == 1:
@@ -77,7 +78,9 @@ class Persistable:
     subsample: None or int, optional, default is None
         Number of datapoints to subsample. The subsample is taken to have a measure
         that approximates the original measure on the full dataset as best as possible,
-        in the Prokhorov sense.
+        in the Prokhorov sense. If metric is ``minkowski`` and the dimensionality is
+        not too big, computing the sample takes time O( log(size_subsample) * size_data ),
+        otherwise it takes time O( size_subsample * size_data ).
 
     n_neighbors: int or string, optional, default is "auto"
         Number of neighbors for each point in X used to initialize
@@ -94,6 +97,10 @@ class Persistable:
         significantly slower because of the GIL, but the backend ``loky`` does
         not work well in some systems.
 
+    n_jobs: int, default is 1
+        Number of processes or threads to use to fit the data structures, for exaple
+        to compute the nearest neighbors of all points in the dataset.
+
     ``**kwargs``:
         Passed to ``KDTree`` or ``BallTree``.
 
@@ -108,14 +115,10 @@ class Persistable:
         n_neighbors="auto",
         debug=False,
         threading=False,
+        n_jobs=4,
         **kwargs
     ):
         self._subsample = None 
-
-        if "leaf_size" in kwargs:
-            leaf_size = kwargs["leaf_size"]
-        else:
-            leaf_size = 40
 
         # if no measure was passed, assume normalized counting measure
         if measure is None:
@@ -129,7 +132,7 @@ class Persistable:
         if subsample is not None:
             if type(subsample) != int:
                 raise ValueError("subsample must be either None or an integer.")
-            ms = _MetricSpace(X, metric, **kwargs)
+            ms = _MetricSpace(X, metric, threading=threading, debug=debug, n_jobs=n_jobs, **kwargs)
             if subsample >= X.shape[0]:
                 subsample = int(X.shape[0] / 4)
                 warnings.warn(
@@ -138,9 +141,14 @@ class Persistable:
                     + " instead."
                 )
             self._subsample = subsample
-            subsample_indices, _, subsample_representatives = ms.close_subsample(
-                subsample
+
+            subsample_euclidean = (metric == "minkowski")
+
+            subsample_indices, subsample_representatives = ms.close_subsample(
+                subsample, euclidean=subsample_euclidean
             )
+            self._subsample = subsample_indices.shape[0]
+
             X = X.copy()
             if metric == "precomputed":
                 X = X[subsample_indices, :][:, subsample_indices]
@@ -149,7 +157,7 @@ class Persistable:
             self._subsample_representatives = subsample_representatives
 
             # compute measure for subsample
-            new_measure = np.zeros(subsample)
+            new_measure = np.zeros(self._subsample)
             for i, _ in enumerate(self._subsample_representatives):
                 new_measure[self._subsample_representatives[i]] += measure[i]
             measure = new_measure
@@ -174,10 +182,11 @@ class Persistable:
             )
         # keep max_k (normalized n_neighbors)
         self._maxk = n_neighbors / X.shape[0]
+        self._dataset_is_large = X.shape[0] > _MANY_POINTS
 
         # construct the filtration
         self._mpspace = _MetricProbabilitySpace(
-            X, metric, measure, n_neighbors, leaf_size, **kwargs
[…]
