### 2023-01-28 b399b47
change zoomed part of component counting function as soon as parameter is changed
