### 2023-01-28 2ac2598
plot PD inside box
