### 2023-01-29 5e8f076
add option to not draw barcode over hilbert function
