### 2023-01-29 55cf250
display vineyards only up to selected point
