### 2023-01-29 43c4525
feat: various interface improvements (#27)
* change zoomed part of component counting function as soon as parameter is changed

* plot PD inside box

* add option to not draw barcode over hilbert function

* display vineyards only up to selected point
