### 2023-01-29 e3825f7
feat: various interface improvements (#27)
* change zoomed part of component counting function as soon as parameter is changed

* plot PD inside box

* add option to not draw barcode over hilbert function

* display vineyards only up to selected point
