+++ persistable/persistable_interactive.py
@@ -52,6 +52,8 @@ CCF_PLOT = "ccf-plot-"
 INTERACTIVE_INPUTS_SELECTION = "interactive-inputs-selection-"
 PV_ENDPOINT_SELECTION = "pv-endpoint-selection-"
 PD_ENDPOINT_SELECTION = "pd-endpoint-selection-"
+PV_DISPLAY_BARCODE = "pv-display-barcode-"
+PD_DISPLAY_BARCODE = "pd-display-barcode-"
 STORED_CCF = "stored-ccf-"
 STORED_X_TICKS_CCF = "stored-x-ticks-ccf-"
 STORED_Y_TICKS_CCF = "stored-y-ticks-ccf-"
@@ -508,45 +510,82 @@ class PersistableInteractive:
                                         ],
                                     ),
                                     html.Div(
-                                        className="parameter-single",
+                                        className="parameters",
                                         id=PV_ENDPOINT_SELECTION_DIV,
                                         children=[
-                                            html.Span(
-                                                className="name",
-                                                children="Endpoint",
-                                            ),
-                                            dcc.RadioItems(
-                                                [
-                                                    "1st line start",
-                                                    "1st line end",
-                                                    "2nd line start",
-                                                    "2nd line end",
-                                                ],
-                                                "1st line start",
-                                                id=PV_ENDPOINT_SELECTION,
-                                                className=VALUE,
-                                            ),
+                                            html.Div(
+                                                className="parameter-single",
+                                                children=[
+                                                    html.Span(
+                                                        className="name",
+                                                        children="Endpoint",
+                                                    ),
+                                                    dcc.RadioItems(
+                                                        [
+                                                            "1st line start",
+                                                            "1st line end",
+                                                            "2nd line start",
+                                                            "2nd line end",
+                                                        ],
+                                                        "1st line start",
+                                                        id=PV_ENDPOINT_SELECTION,
+                                                        className=VALUE,
+                                                    ),
+                                                ]),
+                                            html.Div(
+                                                className="parameter-single",
+                                                children=[
+                                                    html.Span(
+                                                        className="name",
+                                                        children="Display barcode",
+                                                    ),
+                                                    dcc.RadioItems(
+                                                        ["On", "Off"],
+                                                        "Off",
+                                                        id=PV_DISPLAY_BARCODE,
+                                                    ),
+                                                ]
+                                            )
                                         ],
                                     ),
                                     html.Div(
-                                        className="parameter-single",
+                                        className="parameters",
                                         id=PD_ENDPOINT_SELECTION_DIV,
                                         children=[
-                                            html.Span(
-                                                className="name",
-                                                children="Endpoint",
-                                            ),
-                                            dcc.RadioItems(
-                                                [
-                                                    "Line start",
[…]
