### 2023-01-30 ff752a0
add tomato-style flattening
