+++ persistable/persistable_interactive.py
@@ -255,17 +255,19 @@ class PersistableInteractive:
 
             return port
 
-    def cluster(self, tomato_flattening_style=True, keep_low_persistence_clusters=False):
+    def cluster(self, conservative_flattening_style=False, keep_low_persistence_clusters=False):
         """Clusters the dataset with which the Persistable instance that was
         passed through ``run_with`` was initialized.
 
-        tomato_flattening_style: bool, optional, default is True
-            Whether to flatten the hierarchical clustering using the approach
+        conservative_flattening_style: bool, optional, default is False
+            If false, flatten the hierarchical clustering using the approach
             of 'Persistence-Based Clustering in Riemannian Manifolds' Chazal, Guibas,
-            Oudot, Skraba. Otherwise the more conservative and more stable approach of
-            'Stable and consistent density-based clustering' Rolle, Scoccola is used.
+            Oudot, Skraba. If true, use the more conservative and more stable approach
+            of 'Stable and consistent density-based clustering' Rolle, Scoccola is used.
+            The conservative approach usually results in more unclustered points.
 
         keep_low_persistence_clusters: bool, optional, default is False
+            Only relevant if conservative_flattening_style is set to False.
             Whether to keep clusters that are born below the persistence threshold
             associated to the selected n_clusters. If set to True, all points will
             belong to some cluster, but the number of clusters may be larger than the
@@ -285,7 +287,7 @@ class PersistableInteractive:
             )
         else:
             return self._persistable.cluster(**params,
-                tomato_flattening_style = tomato_flattening_style,
+                conservative_flattening_style = conservative_flattening_style,
                 keep_low_persistence_clusters = keep_low_persistence_clusters)
 
     def _chosen_parameters(self):
