### 2023-01-30 854364e
feat: ToMATo-style flattening (#28)
