### 2023-01-30 03cab77
feat: ToMATo-style flattening (#28)
