+++ persistable/persistable.py
@@ -34,6 +34,7 @@ _TOL = 1e-08
 # starting when we consider a dataset large
 _MANY_POINTS = 40000
 
+
 def parallel_computation(function, inputs, n_jobs, debug=False, threading=False):
     if n_jobs == 1:
         return [function(inp) for inp in inputs]
@@ -118,7 +119,7 @@ class Persistable:
         n_jobs=4,
         **kwargs
     ):
-        self._subsample = None 
+        self._subsample = None
 
         # if no measure was passed, assume normalized counting measure
         if measure is None:
@@ -132,7 +133,9 @@ class Persistable:
         if subsample is not None:
             if type(subsample) != int:
                 raise ValueError("subsample must be either None or an integer.")
-            ms = _MetricSpace(X, metric, threading=threading, debug=debug, n_jobs=n_jobs, **kwargs)
+            ms = _MetricSpace(
+                X, metric, threading=threading, debug=debug, n_jobs=n_jobs, **kwargs
+            )
             if subsample >= X.shape[0]:
                 subsample = int(X.shape[0] / 4)
                 warnings.warn(
@@ -142,7 +145,7 @@ class Persistable:
                 )
             self._subsample = subsample
 
-            subsample_euclidean = (metric == "minkowski")
+            subsample_euclidean = metric == "minkowski"
 
             subsample_indices, subsample_representatives = ms.close_subsample(
                 subsample, euclidean=subsample_euclidean
@@ -186,7 +189,14 @@ class Persistable:
 
         # construct the filtration
         self._mpspace = _MetricProbabilitySpace(
-            X, metric, measure, n_neighbors, threading=threading, debug=debug, n_jobs=n_jobs, **kwargs
+            X,
+            metric,
+            measure,
+            n_neighbors,
+            threading=threading,
+            debug=debug,
+            n_jobs=n_jobs,
+            **kwargs
         )
 
         self._bifiltration = _DegreeRipsBifiltration(
@@ -199,9 +209,6 @@ class Persistable:
         self,
         n_neighbors: int = 30,
         n_clusters_range=np.array([3, 15]),
-        propagate_labels=False,
-        n_iterations_propagate_labels=30,
-        n_neighbors_propagate_labels=15,
     ):
         """Clusters the dataset with which the Persistable instance was initialized.
 
@@ -217,17 +224,6 @@ class Persistable:
             A two-element list or tuple representing an integer
             range of possible numbers of clusters to consider when finding the
             optimum number of clusters.
-
-        propagate_labels: bool, optional, default is False
-            Boolean representing whether or not to extend the clustering to
-            noise points by propagating labels.
-
-        n_iterations_propagate_labels: int, optional, default is 30
-            Maximum number of iterations of label propagation to perform. More iterations
-            will cluster more noise points.
-
-        n_neighbors_propagate_labels: int, optional, default is 5
-            How many neighbors to use in the hill climbing procedure.
 
         returns:
             A numpy array of length the number of points in the dataset containing
@@ -259,10 +255,7 @@ class Persistable:
         return self.cluster(
             num_clust,
             [0, k],
-            [s, 0],
-            propagate_labels=propagate_labels,
-            n_iterations_propagate_labels=n_iterations_propagate_labels,
-            n_neighbors_propagate_labels=n_neighbors_propagate_labels,
+            [s, 0]
         )
 
     def cluster(
@@ -270,9 +263,8 @@ class Persistable:
         n_clusters,
         start,
         end,
-        propagate_labels=False,
-        n_iterations_propagate_labels=30,
-        n_neighbors_propagate_labels=15,
+        conservative_flattening_style=False,
+        keep_low_persistence_clusters=False,
     ):
         """Clusters the dataset with which the Persistable instance was initialized.
 
@@ -293,16 +285,19 @@ class Persistable:
[…]
