### 2023-01-30 f7ef33a
feat: ToMATo-style flattening (#28)
