### 2023-01-31 67fda82
add necessary gui functionality
