### 2023-02-02 c3cd6dd
finish dbscan flattening prototype
