### 2023-02-03 0754fbd
change package name to persistable-clustering and add requirements.txt to MANIFEST
