+++ MANIFEST.in
@@ -1,2 +1,2 @@
-include *.md pyproject.toml LICENSE
+include *.md pyproject.toml LICENSE requirements.txt
 recursive-include persistable *.py *.pyx *.pxd *.c *.css *.ico
