### 2023-02-03 c6218e8
add pypi badge and project details to setup.py
