+++ .github/workflows/publish.yaml
@@ -4,28 +4,58 @@ on: [push]
 #  release:
 #    types: [published]
 
-permissions:
-  contents: read
+name: Upload Python Package
 
 jobs:
-  deploy:
+  deploy-unix:
+    strategy:
+      fail-fast: false
 
-    runs-on: ubuntu-latest
-
+    runs-on: "ubuntu-latest"
     steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python
-      uses: actions/setup-python@v3
-      with:
-        python-version: '3.x'
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install build
-    - name: Build package
-      run: python -m build
-    - name: Publish package
-      uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
-      with:
-        user: __token__
-        password: ${{ secrets.PYPI_API_TOKEN }}
+      - uses: actions/checkout@v2
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.x"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install setuptools wheel twine
+      - name: Build and publish
+        env:
+          TWINE_USERNAME: __token__
+          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+        run: |
+          python setup.py sdist
+          twine upload --skip-existing dist/*
+
+#  deploy-windows:
+#    strategy:
+#      fail-fast: false
+#      matrix:
+#        arch: ['x86', 'x64']
+#
+#    runs-on: "windows-latest"
+#
+#    steps:
+#      - uses: actions/checkout@v2
+#      - name: Set up Python
+#        uses: actions/setup-python@v2
+#        with:
+#          architecture: ${{ matrix.arch }}
+#          python-version: "3.x"
+#      - name: Install dependencies
+#        run: |
+#          python -m pip install --upgrade pip
+#          pip install setuptools wheel twine Cython numpy pytest ripser
+#          python setup.py install
+#      - name: Run tests
+#        run: pytest .
+#      - name: Build and publish
+#        env:
+#          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
+#          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
+#        run: |
+#          python setup.py bdist_wheel
+#          twine upload --skip-existing dist/*
