+++ .github/workflows/publish.yaml
@@ -21,6 +21,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install setuptools wheel twine
+          pip install -r requirements.txt
       - name: Build and publish
         env:
           TWINE_USERNAME: __token__
