+++ .github/workflows/run_tests.yaml
@@ -1,4 +1,4 @@
-name: tests
+name: Run Tests
 
 on: [push, pull_request]
 
