+++ .github/workflows/publish.yaml
@@ -1,8 +1,8 @@
 name: Upload Python Package
 
-on: [push]
-#  release:
-#    types: [published]
+on:
+  release:
+    types: [published]
 
 
 jobs:
@@ -29,33 +29,3 @@ jobs:
         run: |
           python setup.py sdist
           twine upload --skip-existing dist/*
-
-#  deploy-windows:
-#    strategy:
-#      fail-fast: false
-#      matrix:
-#        arch: ['x86', 'x64']
-#
-#    runs-on: "windows-latest"
-#
-#    steps:
-#      - uses: actions/checkout@v2
-#      - name: Set up Python
-#        uses: actions/setup-python@v2
-#        with:
-#          architecture: ${{ matrix.arch }}
-#          python-version: "3.x"
-#      - name: Install dependencies
-#        run: |
-#          python -m pip install --upgrade pip
-#          pip install setuptools wheel twine Cython numpy pytest ripser
-#          python setup.py install
-#      - name: Run tests
-#        run: pytest .
-#      - name: Build and publish
-#        env:
-#          TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-#          TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-#        run: |
-#          python setup.py bdist_wheel
-#          twine upload --skip-existing dist/*
