+++ .github/workflows/publish.yaml
@@ -1,4 +1,4 @@
-name: Upload Python Package
+name: upload to pypi
 
 on:
   release:
