+++ .github/workflows/run_tests.yaml
@@ -1,4 +1,4 @@
-name: Run Tests
+name: tests
 
 on: [push, pull_request]
 
