### 2023-02-03 2afe746
feat: add option to extract a single clustering like dbscan does (#29)
