+++ persistable/persistable_interactive.py
@@ -36,6 +36,9 @@ BaseLongCallbackManager.hash_function = monkeypatched_hash_function
 ###
 
 
+X_POINT = "x-point-"
+Y_POINT = "y-point-"
+POINT_SELECTION_DIV = "point-selection-div-"
 X_START_LINE = "x-start-line-"
 Y_START_LINE = "y-start-line-"
 X_END_LINE = "x-end-line-"
@@ -118,6 +121,7 @@ PV_INPUT_GAP = "pv-input-gap-"
 PD_INPUT_GAP = "pd-input-gap-"
 EXPORT_PARAMETERS_BUTTON_PV = "export-parameters-button-pv-"
 EXPORT_PARAMETERS_BUTTON_PD = "export-parameters-button-pd-"
+EXPORT_PARAMETERS_BUTTON_DBSCAN = "export-parameters-button-dbscan-"
 PV_FIXED_PARAMETERS = "fixed-parameters-"
 STORED_PD_BY_PV = "stored-pd-by-pv-"
 STORED_PARAMETERS_AND_PD_BY_PD = "stored-pd-by-pd-"
@@ -258,7 +262,9 @@ class PersistableInteractive:
 
             return port
 
-    def cluster(self, conservative_flattening_style=False, keep_low_persistence_clusters=False):
+    def cluster(
+        self, conservative_flattening_style=False, keep_low_persistence_clusters=False
+    ):
         """Clusters the dataset with which the Persistable instance that was
         passed through ``run_with`` was initialized.
 
@@ -289,9 +295,14 @@ class PersistableInteractive:
                 "No parameters where chosen. Please use the graphical user interface to choose parameters."
             )
         else:
-            return self._persistable.cluster(**params,
-                conservative_flattening_style = conservative_flattening_style,
-                keep_low_persistence_clusters = keep_low_persistence_clusters)
+            if "point" in params:
+                return self._persistable._dbscan_cluster(params["point"])
+            else:
+                return self._persistable.cluster(
+                    **params,
+                    conservative_flattening_style=conservative_flattening_style,
+                    keep_low_persistence_clusters=keep_low_persistence_clusters
+                )
 
     def _chosen_parameters(self):
         self._parameters_sem.acquire()
@@ -309,7 +320,7 @@ class PersistableInteractive:
         # default_s_step = (default_max_s - default_min_s) / 100
         default_granularity_ccf = self._persistable._default_granularity()
         default_granularity_ri = default_granularity_ccf // 5
-        default_granularity_pv =  default_granularity_ccf // 2
+        default_granularity_pv = default_granularity_ccf // 2
         default_num_jobs = 1
         default_max_dim = 15
         default_max_vines = 15
@@ -514,8 +525,9 @@ class PersistableInteractive:
                                             dcc.RadioItems(
                                                 [
                                                     "Off",
-                                                    "One-parameter slice",
-                                                    "Line family",
+                                                    "Single clustering",
+                                                    "Line",
+                                                    "Family of lines",
                                                 ],
                                                 "Off",
                                                 id=INTERACTIVE_INPUTS_SELECTION,
@@ -545,7 +557,8 @@ class PersistableInteractive:
                                                         id=PV_ENDPOINT_SELECTION,
                                                         className=VALUE,
                                                     ),
-                                                ]),
+                                                ],
+                                            ),
                                             html.Div(
                                                 className="parameter-single",
                                                 children=[
@@ -558,8 +571,8 @@ class PersistableInteractive:
                                                         "Off",
[…]
