### 2023-02-03 7998685
reflect installing from pypi in readme
