+++ README.md
@@ -69,18 +69,18 @@ clustering_labels = pi.cluster()
 ## Installing
 
 Make sure you are using Python 3.
-Installation through PyPI is coming soon.
-
-```bash
-pip install git+https://github.com/LuisScoccola/persistable.git
-```
-
-Persistable depends on the following python packages:
+Persistable depends on the following python packages, which will be installed automatically when you install with `pip`:
 `numpy`, `scipy`, `scikit-learn`, `cython`, `plotly`, `dash`, `jupyter_dash`, `diskcache`, `multiprocess`, `psutil`.
 
-## Experimental features
+### Stable
 
-Check the `experimental` branch for more advanced features.
+```bash
+pip install persistable-clustering
+```
+
+### Experimental
+
+The `experimental` branch may contain newer and more advanced features.
 You can install from that branch by running
 
 ```bash
