+++ docs/index.rst
@@ -14,14 +14,22 @@ Installing
 ----------
 
 Make sure you are using Python 3.
-Installation through PyPI is coming soon.
+Persistable depends on the following python packages, which will be installed automatically when you install with `pip`:
+`numpy`, `scipy`, `scikit-learn`, `cython`, `plotly`, `dash`, `jupyter_dash`, `diskcache`, `multiprocess`, `psutil`.
+
+To install the latest release:
 
 .. code-block::
 
-    pip install git+https://github.com/LuisScoccola/persistable.git
+    pip install persistable-clustering
 
-Persistable depends on the following python packages:
-`numpy`, `scipy`, `scikit-learn`, `cython`, `plotly`, `dash`, `jupyter_dash`, `diskcache`, `multiprocess`, `psutil`.
+The `experimental` branch may contain newer and more advanced features.
+You can install from that branch by running:
+
+.. code-block::
+
+    pip install git+https://github.com/LuisScoccola/persistable.git@experimental
+
 
 
 Contents
