### 2023-02-04 22e18f0
change stable to latest release in readme
