### 2023-02-07 LuisScoccola/persistable#31 @AP6YC
PyPi Registration Question
This issue is non-blocking for my JOSS review.

Sorry if I missed it in the documentation somewhere, but is there a plan to register the package on PyPi and perhaps conda-forge? I noticed that the name `persistable` is already taken on PyPi, which is rather unfortunate. Is this the reason, or is something else keeping PyPi registration? Mind you, I don't consider it a big deal to not register on these servers, but there is perhaps added visibility for your package that the convenience for installation could bring.
