### 2023-02-07 LuisScoccola/persistable#32 @AP6YC
Clustering Parameters Question
This question is non-blocking for my JOSS review.

Is there a plan to make the parameters stateful or transferrable? For example, after I select some desirable parameters using the `persistable` method, where do the parameters "live," so to speak, and how could I use these parameters to cluster on another dataset after reloading? I do understand that there is a `quick_cluster` method in your API, which I appreciate, but as I understand it this method automatically selects parameters based upon the ranges provided. Sorry again if this is due to my misunderstanding of the tool, which is why I add this issue as non-blocking.
