### 2023-02-07 LuisScoccola/persistable#33 @AP6YC
API Index
This is a relatively low-level request that I have for the documentation, which I otherwise enjoy, but I would like to see an index of the core types and methods of the package in the API reference section. For example, I wonder if it is possible in the `sphinx` docs setup that you use for the documentation to, in the API section, have a list of types and their methods that link to their full definitions further down the page (in the way that you have currently).

If this is limited by the tools that `sphinx` provides, it is no worries. I just know that other packages such as those for documenting Julia code allow for an index functionality.
