### 2023-02-07 #5022 @AP6YC
I am overall happy with the contributions of the package, and I would like to congratulate the authors and programmers for making such useful and relevant software. The authors should feel proud of their work.

I apologize for the delay, but I have nearly completed my review, contingent on only a few points. I have provided two non-blocking issues on the repo in the form of questions and suggestions, and I have one minor blocking issue for improving the API reference. In short, I have a suggestion for the improvement of the API index that I would like addressed (with hopefully minimal time required of the authors) and two questions that may be long-term changes, which is why I do not wish for them to be addressed before my approval.

These issues are:

Non-blocking (not necessary to complete my review):
- https://github.com/LuisScoccola/persistable/issues/31
- https://github.com/LuisScoccola/persistable/issues/32

Blocking (small suggestions that I would like implemented or explained):
- https://github.com/LuisScoccola/persistable/issues/33
