### 2023-02-07 LuisScoccola/persistable#31 @LuisScoccola
Thank you for bringing this up! There is such a plan: the package is currently registered as [persistable-clustering](https://pypi.org/project/persistable-clustering/). This will be added to the readme and documentation with the next release, which will be done in the next week or so.
