### 2023-02-07 LuisScoccola/persistable#32 @LuisScoccola
Thank you for pointing this out. The parameters selected with the UI can be accessed by calling `PersistableInteractive`'s method `_chosen_parameters()`. The parameters can then be used as input for `Persistable`'s method `cluster()`. However, right now `_chosen_parameters()`  is an internal method and is not documented. I think it is a good idea to turn it into a user-facing method and to add it to the API. This will be done in the next release.
