### 2023-02-07 LuisScoccola/persistable#33 @LuisScoccola
Thank you for the suggestion. I think I understand what you are referring to, but, just to be sure, could you please provide an example?
