+++ docs/api.rst
@@ -1,6 +1,34 @@
 API Reference
 =============
 
+.. rubric:: Classes
+
+.. autosummary::
+    :nosignatures:
+
+    persistable.Persistable
+    persistable.PersistableInteractive
+
+.. rubric:: persistable.Persistable
+
+.. autosummary::
+    :nosignatures:
+
+    persistable.Persistable.cluster
+    persistable.Persistable.quick_cluster
+
+.. rubric:: persistable.PersistableInteractive
+
+.. autosummary::
+    :nosignatures:
+
+    persistable.PersistableInteractive.cluster
+    persistable.PersistableInteractive.start_UI
+
+
+Details
+-------
+
 .. autoclass:: persistable.Persistable
     :members:
 
