### 2023-02-07 2eb830d
improve docstrings
