### 2023-02-07 LuisScoccola/persistable#33 @AP6YC
Sure thing! I am thinking of the following from the Julia `Documenter.jl` package (which is actually itself used for generating documentation for Julia projects, like `sphinx` is frequently used):

https://documenter.juliadocs.org/stable/lib/public/

It may just come in the form of the `toctree` directive (sphinx docs): https://www.sphinx-doc.org/en/master/usage/restructuredtext/directives.html#directive-toctree
Or perhaps an autosummary (sphinx rtd theme docs): https://sphinx-rtd-theme.readthedocs.io/en/stable/demo/api.html#generated-index
Or a module index with links (example): https://documen.tician.de/codepy/py-modindex.html

But I also see that it may be comparing apples to oranges. After more digging, I don't see an obvious way of doing this in the Python ecosystem. I don't want to be to picky, and defer to what you think is best for the project.
