### 2023-02-07 159f749
other small improvements to docstring
