+++ persistable/persistable.py
@@ -23,7 +23,7 @@ _TOL = 1e-08
 
 
 class Persistable:
-    """Does density-based clustering on finite metric spaces.
+    """Density-based clustering on finite metric spaces.
 
     Persistable has two main clustering methods: ``cluster()`` and ``quick_cluster()``.
     The methods are similar, the main difference being that ``quick_cluster()`` takes
@@ -192,7 +192,7 @@ class Persistable:
         n_iterations_extend_cluster=10,
         n_neighbors_extend_cluster=5,
     ):
-        """Clusters dataset passed at initialization.
+        """Cluster dataset passed at initialization.
 
         n_clusters: int
             Integer determining how many clusters the final clustering
