### 2023-02-07 d19f6ab
docs: add index to API Reference (#34)
Address #33

* add index to api in docs

* minor improvements to docstrings
