+++ docs/api.rst
@@ -1,6 +1,37 @@
 API Reference
 =============
 
+Index
+-----
+
+.. rubric:: Classes
+
+.. autosummary::
+    :nosignatures:
+
+    persistable.Persistable
+    persistable.PersistableInteractive
+
+.. rubric:: persistable.Persistable methods
+
+.. autosummary::
+    :nosignatures:
+
+    persistable.Persistable.cluster
+    persistable.Persistable.quick_cluster
+
+.. rubric:: persistable.PersistableInteractive methods
+
+.. autosummary::
+    :nosignatures:
+
+    persistable.PersistableInteractive.cluster
+    persistable.PersistableInteractive.start_UI
+
+
+Details
+-------
+
 .. autoclass:: persistable.Persistable
     :members:
 
