### 2023-02-07 LuisScoccola/persistable#33 @LuisScoccola
Thank you for the clarifications. Here's an attempt: https://persistable.readthedocs.io/en/latest/api.html
