+++ .github/workflows/publish.yaml
@@ -2,15 +2,40 @@ name: upload to pypi
 
 on:
   release:
-    types: [published]
-
+    types: [push]
 
 jobs:
   deploy-unix:
     strategy:
       fail-fast: false
 
-    runs-on: "ubuntu-latest"
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.x"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install setuptools wheel twine
+          pip install -r requirements.txt
+      - name: Build and publish
+        env:
+          TWINE_USERNAME: __token__
+          TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
+        run: |
+          python setup.py sdist bdist_wheel
+          twine upload dist/*.tar.gz
+          auditwheel repair dist/*linux_x86_64.whl
+          twine upload wheelhouse/*.whljobs:
+
+  deploy-windows:
+    strategy:
+      fail-fast: false
+
+    runs-on: windows-2019
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python
@@ -28,4 +53,28 @@ jobs:
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
           python setup.py sdist
-          twine upload --skip-existing dist/*
+          twine upload dist/*
+
+  deploy-macos:
+    strategy:
+      fail-fast: false
+
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: "3.x"
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install setuptools wheel twine
+          pip install -r requirements.txt
+      - name: Build and publish
+        env:
+          TWINE_USERNAME: __token__
+          TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
+        run: |
+          python setup.py sdist
+          twine upload dist/*
