+++ .github/workflows/publish.yaml
@@ -1,8 +1,6 @@
 name: upload to pypi
 
-on:
-  release:
-    types: [push]
+on: [push]
 
 jobs:
   deploy-unix:
