+++ .github/workflows/publish.yaml
@@ -27,7 +27,7 @@ jobs:
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
           python setup.py sdist bdist_wheel
-          twine upload dist/*.tar.gz
+          twine upload --skip-existing dist/*.tar.gz
           auditwheel repair dist/*linux_x86_64.whl
           twine upload --skip-existing wheelhouse/*.whljobs:
 
@@ -52,7 +52,7 @@ jobs:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
-          python setup.py sdist
+          python setup.py sdist bdist_wheel
           twine upload --skip-existing dist/*
 
   deploy-macos:
@@ -76,5 +76,5 @@ jobs:
           TWINE_USERNAME: __token__
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
-          python setup.py sdist
+          python setup.py sdist bdist_wheel
           twine upload --skip-existing dist/*
