+++ .github/workflows/publish.yaml
@@ -19,7 +19,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools wheel twine
+          pip install setuptools wheel twine auditwheel
           pip install -r requirements.txt
       - name: Build and publish
         env:
