+++ .github/workflows/publish.yaml
@@ -27,9 +27,7 @@ jobs:
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
           python setup.py sdist bdist_wheel
-          twine upload --skip-existing dist/*.tar.gz
-          auditwheel repair dist/*linux_x86_64.whl
-          twine upload --skip-existing wheelhouse/*.whljobs:
+          twine upload --skip-existing dist/*
 
   deploy-windows:
     strategy:
