+++ .github/workflows/publish.yaml
@@ -27,7 +27,9 @@ jobs:
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
           python setup.py sdist bdist_wheel
-          twine upload --skip-existing dist/*
+          twine upload --skip-existing dist/*.tar.gz
+          auditwheel repair --plat manylinux_2_30_x86_64 dist/*linux_x86_64.whl
+          twine upload --skip-existing wheelhouse/*.whljobs:
 
   deploy-windows:
     strategy:
