+++ .github/workflows/publish.yaml
@@ -19,7 +19,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools wheel twine auditwheel
+          pip install setuptools wheel twine auditwheel patchelf
           pip install -r requirements.txt
       - name: Build and publish
         env:
