+++ .github/workflows/publish.yaml
@@ -29,7 +29,7 @@ jobs:
           python setup.py sdist bdist_wheel
           twine upload --skip-existing dist/*.tar.gz
           auditwheel repair --plat manylinux_2_17_x86_64 dist/*linux_x86_64.whl
-          twine upload --skip-existing wheelhouse/*.whljobs:
+          twine upload --skip-existing dist/*
 
   deploy-windows:
     strategy:
