+++ .github/workflows/publish.yaml
@@ -28,8 +28,8 @@ jobs:
         run: |
           python setup.py sdist bdist_wheel
           twine upload --skip-existing dist/*.tar.gz
-          auditwheel repair --plat manylinux_2_17_x86_64 dist/*linux_x86_64.whl
-          twine upload --skip-existing dist/*
+          auditwheel repair --plat manylinux_2_17_x86_64 dist/*linux_x86_64.whl -w dist1
+          twine upload --skip-existing dist1/*
 
   deploy-windows:
     strategy:
