+++ .github/workflows/publish.yaml
@@ -59,7 +59,7 @@ jobs:
     strategy:
       fail-fast: false
 
-    runs-on: macos-latest
+    runs-on: macos-11
     steps:
       - uses: actions/checkout@v2
       - name: Set up Python
