### 2023-02-22 11653a4
ci: add push to pypi on release (#35)
