### 2023-03-03 e8f09cf
add joss badge and bump version
