### 2023-03-03 4bea5e0
chore: add joss badge and bump version (#36)
