+++ persistable/persistable_interactive.py
@@ -311,45 +311,46 @@ class PersistableInteractive:
         return params
 
     def _layout_gui(self):
-        default_min_k = 0
-        end = self._persistable._find_end()
-        default_max_k = end[1]
-        # default_k_step = default_max_k / 100
-        default_min_s = 0
-        default_max_s = end[0]
-        # default_s_step = (default_max_s - default_min_s) / 100
-        default_granularity_ccf = self._persistable._default_granularity()
-        default_granularity_ri = default_granularity_ccf // 5
-        default_granularity_pv = default_granularity_ccf // 2
-        default_num_jobs = 1
-        default_max_dim = 15
-        default_max_vines = 15
-        min_granularity = 2
-        max_granularity = 512
-        max_granularity_ri = 64
-        min_granularity_vineyard = 1
-        max_granularity_vineyard = max_granularity
-        defr = 6
-        default_x_start_first_line = (default_min_s + default_max_s) * (1 / defr)
-        default_y_start_first_line = (default_min_k + default_max_k) * (1 / 2)
-        default_x_end_first_line = (default_max_s + default_min_s) * (1 / 2)
-        default_y_end_first_line = (default_min_k + default_max_k) * (1 / defr)
-        default_x_start_second_line = (default_min_s + default_max_s) * (1 / 2)
-        default_y_start_second_line = (default_min_k + default_max_k) * (
-            (defr - 1) / defr
-        )
-        default_x_end_second_line = (default_max_s + default_min_s) * (
-            (defr - 1) / defr
-        )
-        default_y_end_second_line = (default_min_k + default_max_k) * (1 / 2)
-        default_x_start_line = (
-            default_x_start_first_line + default_x_start_second_line
-        ) / 2
-        default_y_start_line = (
-            default_y_start_first_line + default_y_start_second_line
-        ) / 2
-        default_x_end_line = (default_x_end_first_line + default_x_end_second_line) / 2
-        default_y_end_line = (default_y_end_first_line + default_y_end_second_line) / 2
+
+        def compute_defaults(end, default_granularity):
+            d0 = {"granularity": default_granularity}
+            d1 = {
+                "min_k": 0,
+                "max_k": end[1],
+                "min_s": 0,
+                "max_s": end[0],
+                "granularity_ri": d0["granularity"] // 5,
+                "granularity_pv": d0["granularity"] // 2,
+                "num_jobs": 1,
+                "max_dim": 15,
+                "max_vines" : 15,
+                "min_granularity" : 2,
+                "max_granularity" : 512,
+                "max_granularity_ri" : 64,
+                "min_granularity_vineyard" : 1,
+                "defr" : 6,
+            }
+            d2 = {
+                "max_granularity_vineyard" : d1["max_granularity"],
+                "x_start_first_line" : (d1["min_s"] + d1["max_s"]) * (1 / d1["defr"]),
+                "y_start_first_line" : (d1["min_k"] + d1["max_k"]) * (1 / 2),
+                "x_end_first_line" : (d1["max_s"] + d1["min_s"]) * (1 / 2),
+                "y_end_first_line" : (d1["min_k"] + d1["max_k"]) * (1 / d1["defr"]),
+                "x_start_second_line" : (d1["min_s"] + d1["max_s"]) * (1 / 2),
+                "y_start_second_line" : (d1["min_k"] + d1["max_k"]) * ( (d1["defr"] - 1) / d1["defr"]),
+                "x_end_second_line" : (d1["max_s"] + d1["min_s"]) * ( (d1["defr"] - 1) / d1["defr"]),
[…]
