### 2023-03-03 0e4b8d6
update html labels in ui tests
