### 2023-03-03 54d855c
fix problem when copying chosen parameters is none
