### 2023-03-06 9e5f195
fix typo in error message
