### 2023-03-07 86cb982
add covariance and other features to saving function
