+++ persistable/persistable_interactive.py
@@ -2957,6 +2957,10 @@ class PersistableInteractive:
                 [NUM_JOBS_CCF, VALUE, IN],
                 [NUM_JOBS_PV, VALUE, IN],
                 [NUM_JOBS_RI, VALUE, IN],
+                [Y_COVARIANT, VALUE, IN],
+                [LINE, VALUE, IN],
+                [PV_GAP, VALUE, IN],
+                [PD_GAP, VALUE, IN],
                 [MAX_COMPONENTS, VALUE, IN],
                 [MAX_VINES, VALUE, IN],
                 [X_START_FIRST_LINE, VALUE, IN],
