### 2023-03-07 7a999f4
feat: add subsampling (#38)
* add description of tests

* refactor mpspace and use metric space, metric probability space, and degree-Rips bifiltration

* simplify shifts in displaying of hilbert function and rank decomposition

* add todo for persistent metric space

* add naive subsampling for fast metrics and distance matrix

* add subsampling option to user-facing persistable class

* add measure input description to persistable init
