+++ persistable/persistable.py
@@ -10,6 +10,7 @@ from .borrowed.prim_mst import mst_linkage_core_vector
 from .borrowed.dense_mst import stepwise_dendrogram_with_core_distances
 from .borrowed.dist_metrics import DistanceMetric
 from .auxiliary import lazy_intersection
+from .subsampling import close_subsample_fast_metric, close_subsample_distance_matrix
 from .persistence_diagram_h0 import persistence_diagram_h0
 from .signed_betti_numbers import (
     signed_betti,
@@ -67,6 +68,17 @@ class Persistable:
         ``from sklearn.neighbors import KDTree, BallTree``) or ``"precomputed"``
         if X is a distance matrix.
 
+    measure: None or ndarray(n_samples), default is None
+        A numpy vector of length (samples) of non-negative numbers, which is
+        intepreted as a measure on the data points.
+        If None, the uniform measure where each point has weight 1/samples is used.
+        If the measure does not sum to 1, it is normalized.
+
+    subsample: None or int, optional, default is None
+        Number of datapoints to subsample. The subsample is taken to have a measure
+        that approximates the original measure on the full dataset as best as possible,
+        in the Prokhorov sense.
+
     n_neighbors: int or string, optional, default is "auto"
         Number of neighbors for each point in X used to initialize
         datastructures used for clustering. If set to ``"all"`` it will use
@@ -91,57 +103,93 @@ class Persistable:
         self,
         X,
         metric="minkowski",
+        measure=None,
+        subsample=None,
         n_neighbors="auto",
         debug=False,
         threading=False,
         **kwargs
     ):
-        self._debug = debug
-        self._threading = threading
-        # keep dataset
-        self._data = X
-        # if metric is minkowski but no p was passed, assume p = 2
-        if metric == "minkowski" and "p" not in kwargs:
-            kwargs["p"] = 2
-        # if no measure was passed, assume normalized counting measure
-        if "measure" not in kwargs:
-            measure = np.full(X.shape[0], 1.0 / X.shape[0])
-        elif "measure" in kwargs and kwargs["measure"] is None:
-            measure = np.full(X.shape[0], 1.0 / X.shape[0])
-            del kwargs["measure"]
-        else:
-            measure = kwargs["measure"]
-            del kwargs["measure"]
+        self._subsample = None 
+
         if "leaf_size" in kwargs:
             leaf_size = kwargs["leaf_size"]
         else:
             leaf_size = 40
-        self._mpspace = _MetricProbabilitySpace(
-            X, metric, measure, leaf_size, debug=debug, threading=threading, **kwargs
-        )
+
+        # if no measure was passed, assume normalized counting measure
+        if measure is None:
+            measure = np.full(X.shape[0], 1.0 / X.shape[0])
+        else:
+            measure = np.array(measure)
+            assert np.all(measure >= 0)
+            # rescale so that total measure is 1
+            measure = measure / np.sum(measure)
+
+        if subsample is not None:
+            if type(subsample) != int:
+                raise ValueError("subsample must be either None or an integer.")
+            ms = _MetricSpace(X, metric, **kwargs)
+            if subsample >= X.shape[0]:
+                subsample = int(X.shape[0] / 4)
+                warnings.warn(
+                    "subsample is greater than or equal to number of datapoints, using "
+                    + str(subsample)
+                    + " instead."
+                )
+            self._subsample = subsample
+            subsample_indices, _, subsample_representatives = ms.close_subsample(
+                subsample
+            )
+            X = X.copy()
+            if metric == "precomputed":
[…]
