### 2023-03-07 1bfebfd
feat: add parameter selection using a single line (#40)
* add toggle for one-parameter slice selection

* add timeout to ui tests so that windows tests don't fail sometimes

* add inputs for persistence diagram choice

* add interactivity to one-parameter slice selection

* add persistence diagram window
