### 2023-03-07 269add4
perf: use two different metric trees for nn and mst, and add faster subsampling for Euclidean data (#41)
* use different trees for nn_fit and for spanning tree computation

* better defaults for granularity for large data

* add fast subsampling for euclidean data
