### 2023-03-07 ce94745
feat: save session (#44)
* add save session feature

* reflect interface change in tests
