+++ persistable/persistable_interactive.py
@@ -81,18 +81,23 @@ COMPUTE_CCF_BUTTON = "compute-ccf-button-"
 STOP_COMPUTE_CCF_BUTTON = "stop-compute-ccf-button-"
 COMPUTE_RI_BUTTON = "compute-ri-button-"
 STOP_COMPUTE_RI_BUTTON = "stop-compute-ri-button-"
-INPUT_GRANULARITY_CCF = "input-granularity-ccf-"
-INPUT_GRANULARITY_RI = "input-granularity-ri-"
-INPUT_NUM_JOBS_CCF = "input-num-jobs-ccf-"
-INPUT_NUM_JOBS_RI = "input-num-jobs-ri-"
-INPUT_MAX_COMPONENTS = "input-max-components-"
-INPUT_MAX_RI = "input-max-ri-"
-INPUT_MIN_LENGTH_RI = "input-min-length-bars-ri-"
-INPUT_SIGNED_BETTI_NUMBERS = "input-signed-betti-numbers-"
-INPUT_Y_COVARIANT = "input-y-covariant-"
-INPUT_DISPLAY_RI = "input-display-ri-"
-INPUT_DECOMPOSE_BY_RI = "input-decompose-by-ri-"
-INPUT_REDUCED_HOMOLOGY_RI = "input-reduced-homology-ri-"
+GRANULARITY = "granularity-"
+MIN_GRANULARITY = "min-granularity-"
+MAX_GRANULARITY = "max-granularity-"
+MAX_GRANULARITY_RI = "max-granularity-ri-"
+MIN_GRANULARITY_VINEYARD = "min-granularity-vineyard-"
+MAX_GRANULARITY_VINEYARD = "max-granularity-vineyard-"
+GRANULARITY_RI = "granularity-ri-"
+NUM_JOBS_CCF = "num-jobs-ccf-"
+NUM_JOBS_RI = "num-jobs-ri-"
+MAX_COMPONENTS = "max-components-"
+MAX_RI = "max-ri-"
+MIN_LENGTH_RI = "min-length-bars-ri-"
+SIGNED_BETTI_NUMBERS = "signed-betti-numbers-"
+Y_COVARIANT = "y-covariant-"
+DISPLAY_RI = "display-ri-"
+DECOMPOSE_BY_RI = "decompose-by-ri-"
+REDUCED_HOMOLOGY_RI = "reduced-homology-ri-"
 CCF_PLOT_CONTROLS_DIV = "ccf-plot-controls-div-"
 CCF_DETAILS = "ccf-details-"
 CCF_EXTRAS = "ccf-extras-"
@@ -105,8 +110,8 @@ PD_PANEL = "pd-panel-"
 LOG = "log-"
 LOG_DIV = "log-div-"
 STORED_PV = "stored-pv-"
-INPUT_MAX_VINES = "input-max-vines-"
-INPUT_PROM_VIN_SCALE = "input-prom-vin-scale-"
+MAX_VINES = "max-vines-"
+PROM_VIN_SCALE = "prom-vin-scale-"
 COMPUTE_PV_BUTTON = "compute-pv-button-"
 STOP_COMPUTE_PV_BUTTON = "stop-compute-pv-button-"
 COMPUTE_PD_BUTTON = "compute-pd-button-"
@@ -114,11 +119,11 @@ STOP_COMPUTE_PD_BUTTON = "stop-compute-pd-button-"
 PV_PLOT = "pv-plot-"
 PD_PLOT = "pd-plot-"
 STORED_PV_DRAWING = "stored-pv-drawing-"
-INPUT_GRANULARITY_PV = "input-granularity-pv-"
[…]
