### 2023-03-07 9d63b52
doc: improve persistableInteractive cluster docstring
