### 2023-03-07 19985e3
revert: hide rank decomposition
