### 2023-03-08 c720278
docs: add bibtex entry to readme
